### 2020-08-31 #2623 @whedon
[PRE REVIEW]: SLIX: A Python package for fully automated evaluation of Scattered Light Imaging measurements on brain tissue
**Submitting author:** @Thyre (<a href="http://orcid.org/0000-0002-1219-0310">Jan André Reuter</a>)
**Repository:** <a href="https://github.com/3d-pli/SLIX" target ="_blank">https://github.com/3d-pli/SLIX</a>
**Version:** v1.1
