### 2020-09-04 2983f1d
Readded test file after it got lost somehow in the last commit
