### 2020-09-11 #2623 @danielskatz
@Thyre - as I look for an editor, who will then look for reviewers, I wanted to share a comment that we had during our scope discussion: "Looks like it should be added on PyPI and turned into a callable library (that could also offer a CLI)"  This is just input, not a requirement.
