### 2020-09-25 9d9a80c
Initial commit. Number of peaks and first version of prominence peaks can be extracted. Prominence maps isn't the same as the one in SLIX yet because the first and last measurements are not yet considered for prominence calculation. This will be changed soon.
Signed-off-by: Jan Andre Reuter <jan.andre.reuter@hotmail.de>
