+++ SLIX/toolbox.py
@@ -1,0 +1,169 @@
+import numpy
+import cupy
+import nibabel
+import tifffile
+from numba import cuda
+
+# DEFAULT PARAMETERS
+BACKGROUND_COLOR = -1
+MAX_DISTANCE_FOR_CENTROID_ESTIMATION = 2
+
+NUMBER_OF_SAMPLES = 100
+TARGET_PEAK_HEIGHT = 0.94
+TARGET_PROMINENCE = 0.08
+
+
+def read_image(FILEPATH):
+    """
+    Reads image file and returns it.
+    Supported file formats: NIfTI, Tiff.
+
+    Arguments:
+        FILEPATH: Path to image
+
+    Returns:
+        numpy.array: Image with shape [x, y, z] where [x, y] is the size of a single image and z specifies the number
+                     of measurements
+    """
+    # Load NIfTI dataset
+    if FILEPATH.endswith('.nii'):
+        data = nibabel.load(FILEPATH).get_fdata()
+        data = numpy.squeeze(numpy.swapaxes(data, 0, 1))
+    else:
+        data = tifffile.imread(FILEPATH)
+        data = numpy.squeeze(numpy.moveaxis(data, 0, -1))
+
+    return data
+
+
+def peaks(image, return_numpy=True):
+    gpu_image = cupy.array(image, dtype='float32')
+    right = cupy.roll(gpu_image, 1, axis=-1) - gpu_image
+    left = cupy.roll(gpu_image, -1, axis=-1) - gpu_image
+    del gpu_image
+
+    peaks = (left < 0) & (right <= 0)
+    del right
+    del left
+
+    if return_numpy:
+        peaks_cpu = cupy.asnumpy(peaks)
+        del peaks
+
+        return peaks_cpu
+    else:
+        return peaks
+
+
+def num_peaks(image, return_numpy=True):
+    gpu_image = cupy.array(image, dtype='float64')
+    right = cupy.roll(gpu_image, 1, axis=-1) - gpu_image
+    left = cupy.roll(gpu_image, -1, axis=-1) - gpu_image
+    del gpu_image
+
+    peaks = (left < 0) & (right <= 0)
+    del right
+    del left
+
+    resulting_image = cupy.empty((peaks.shape[:2]))
+    resulting_image[:, :] = cupy.count_nonzero(peaks, axis=-1)
+    if return_numpy:
+        resulting_image_cpu = cupy.asnumpy(resulting_image)
+        del resulting_image
+        return resulting_image_cpu
+    else:
+        return resulting_image
+
+
+def normalize(image, kind_of_normalization=0, return_numpy=True):
+    gpu_image = cupy.array(image, dtype='float32')
+    if kind_of_normalization == 0:
+        gpu_image = (gpu_image - cupy.min(gpu_image, axis=-1)[..., None]) / \
+                    (cupy.max(gpu_image, axis=-1)[..., None] - cupy.min(gpu_image, axis=-1)[..., None])
+    else:
+        gpu_image = gpu_image / cupy.mean(gpu_image)
+
+    if return_numpy:
+        gpu_image_cpu = cupy.asnumpy(gpu_image)
+        del gpu_image
+        return gpu_image_cpu
+    else:
+        return gpu_image
+
+
+@cuda.jit()
+def _prominence(image, peak_array, result_image):
+    idx = cuda.grid(1)
+    sub_image = image[idx]
+    sub_peak_array = peak_array[idx]
+    for pos, entry in enumerate(sub_peak_array):
+        if entry >= 0:
+            i_min = 0
+            i_max = len(sub_peak_array) - 1
+
+            i = pos
+            left_min = sub_image[pos]
+            while i_min <= i and sub_image[i] <= sub_image[pos]:
+                if sub_image[i] < left_min:
+                    left_min = sub_image[i]
[…]
