### 2020-09-25 e6f0eaa
Fixed and finished prominence calculation. Fixed a bug in normalization method where the whole image was used to normalize the line profiles instead of the individual line profiles. Prominence calculation will only respect peaks with a high prominence for now. Please keep that in mind.
Signed-off-by: Jan Andre Reuter <jan.andre.reuter@hotmail.de>
