### 2020-09-25 c35f80c
Direction implemented.
The implementation in its current state is way faster than the current CPU implementation. The old version should not be considered once this version is complete.

Signed-off-by: Jan Andre Reuter <jan.andre.reuter@hotmail.de>
