+++ SLIX/toolbox.py
@@ -4,6 +4,8 @@ import nibabel
 import tifffile
 from numba import cuda
 
+from matplotlib import pyplot as plt
+
 # DEFAULT PARAMETERS
 BACKGROUND_COLOR = -1
 MAX_DISTANCE_FOR_CENTROID_ESTIMATION = 2
@@ -103,10 +105,10 @@ def _prominence(image, peak_array, result_image):
             i_max = int(len(sub_peak_array) * 1.5)
 
             i = pos
-            left_min = image[idx, pos]
-            while i_min <= i and image[idx, i] <= image[idx, pos]:
-                if image[idx, i] < left_min:
-                    left_min = image[idx, i]
+            left_min = sub_image[pos]
+            while i_min <= i and sub_image[i] <= sub_image[pos]:
+                if sub_image[i] < left_min:
+                    left_min = sub_image[i]
                 i -= 1
 
             i = pos
@@ -121,6 +123,43 @@ def _prominence(image, peak_array, result_image):
             result_image[idx, pos] = 0
 
 
+@cuda.jit()
+def _peakwidth(image, peak_array, result_image):
+    pass
+
+
+@cuda.jit('void(int8[:, :], int8[:], float32[:, :])')
+def _direction(peak_array, number_of_peaks, result_image):
+    idx = cuda.grid(1)
+    sub_peak_array = peak_array[idx]
+    num_directions = result_image.shape[-1]
+
+    current_direction = 0
+
+    result_image[idx, :] = BACKGROUND_COLOR
+    if number_of_peaks[idx] // 2 <= num_directions:
+        for i in range(len(sub_peak_array)):
+            if sub_peak_array[i] == 1:
+                left = i * 360.0 / len(sub_peak_array)
+                if number_of_peaks[idx] == 1:
+                    result_image[idx, current_direction] = (270.0 - left) % 180
+                    break
+                elif number_of_peaks[idx] % 2 == 0:
+                    right_side_peak = number_of_peaks[idx]//2
+                    current_position = i+1
+                    while right_side_peak > 0 and current_position < len(sub_peak_array):
+                        if sub_peak_array[current_position] == 1:
+                            right_side_peak = right_side_peak - 1
+                        current_position = current_position + 1
+                    right = (current_position-1) * 360.0 / len(sub_peak_array)
+                    if number_of_peaks[idx] == 2 or abs(180 - (right - left)) < 35:
+                        result_image[idx, current_direction] = (270.0 - ((left + right) / 2.0)) % 180
+                    current_direction += 1
+
+                if current_direction == number_of_peaks[idx]//2:
+                    break
+
+
 def peak_prominence(image, peak_image=None, kind_of_normalization=0, return_numpy=True):
     gpu_image = cupy.array(image, dtype='float32')
     if peak_image is not None:
@@ -133,15 +172,13 @@ def peak_prominence(image, peak_image=None, kind_of_normalization=0, return_nump
 
     gpu_image = gpu_image.reshape(image_x * image_y, image_z)
     gpu_peak_image = gpu_peak_image.reshape(image_x * image_y, image_z).astype('int8')
-    result_img = numpy.zeros((image_x * image_y, image_z), dtype='float32')
-    result_img_gpu = cuda.to_device(result_img)
-    del result_img
+    result_img_gpu = cupy.empty((image_x * image_y, image_z), dtype='float32')
 
     # https://github.com/scipy/scipy/blob/master/scipy/signal/_peak_finding_utils.pyx
     threadsperblock = 256
[…]
