+++ SLIX/toolbox.py
@@ -1,38 +1,30 @@
 import numpy
 import cupy
-import nibabel
-import tifffile
 from numba import cuda
 
-from matplotlib import pyplot as plt
-
-# DEFAULT PARAMETERS
-BACKGROUND_COLOR = -1
-MAX_DISTANCE_FOR_CENTROID_ESTIMATION = 2
-
-NUMBER_OF_SAMPLES = 100
-TARGET_PEAK_HEIGHT = 0.94
-TARGET_PROMINENCE = 0.08
+import nibabel
+import tifffile
+from SLIX._toolbox import _direction, _prominence, _peakwidth
 
 
-def read_image(FILEPATH):
+def read_image(filepath):
     """
     Reads image file and returns it.
     Supported file formats: NIfTI, Tiff.
 
     Arguments:
-        FILEPATH: Path to image
+        filepath: Path to image
 
     Returns:
         numpy.array: Image with shape [x, y, z] where [x, y] is the size of a single image and z specifies the number
                      of measurements
     """
     # Load NIfTI dataset
-    if FILEPATH.endswith('.nii'):
-        data = nibabel.load(FILEPATH).get_fdata()
+    if filepath.endswith('.nii'):
+        data = nibabel.load(filepath).get_fdata()
         data = numpy.squeeze(numpy.swapaxes(data, 0, 1))
     else:
-        data = tifffile.imread(FILEPATH)
+        data = tifffile.imread(filepath)
         data = numpy.squeeze(numpy.moveaxis(data, 0, -1))
 
     return data
@@ -93,72 +85,6 @@ def normalize(image, kind_of_normalization=0, return_numpy=True):
         return gpu_image
 
 
-@cuda.jit('void(float32[:, :], int8[:, :], float32[:, :])')
-def _prominence(image, peak_array, result_image):
-    idx = cuda.grid(1)
-    sub_image = image[idx]
-    sub_peak_array = peak_array[idx]
-
-    for pos in range(len(sub_peak_array)):
-        if sub_peak_array[pos] == 1:
-            i_min = -len(sub_peak_array) / 2
-            i_max = int(len(sub_peak_array) * 1.5)
-
-            i = pos
-            left_min = sub_image[pos]
-            while i_min <= i and sub_image[i] <= sub_image[pos]:
-                if sub_image[i] < left_min:
-                    left_min = sub_image[i]
-                i -= 1
-
-            i = pos
-            right_min = sub_image[pos]
-            while i <= i_max and sub_image[i % len(sub_peak_array)] <= sub_image[pos]:
-                if sub_image[i % len(sub_peak_array)] < right_min:
-                    right_min = sub_image[i % len(sub_peak_array)]
-                i += 1
-
-            result_image[idx, pos] = sub_image[pos] - max(left_min, right_min)
-        else:
-            result_image[idx, pos] = 0
-
-
-@cuda.jit()
-def _peakwidth(image, peak_array, result_image):
-    pass
-
-
-@cuda.jit('void(int8[:, :], int8[:], float32[:, :])')
-def _direction(peak_array, number_of_peaks, result_image):
-    idx = cuda.grid(1)
-    sub_peak_array = peak_array[idx]
-    num_directions = result_image.shape[-1]
-
-    current_direction = 0
-
-    result_image[idx, :] = BACKGROUND_COLOR
-    if number_of_peaks[idx] // 2 <= num_directions:
-        for i in range(len(sub_peak_array)):
-            if sub_peak_array[i] == 1:
-                left = i * 360.0 / len(sub_peak_array)
-                if number_of_peaks[idx] == 1:
-                    result_image[idx, current_direction] = (270.0 - left) % 180
-                    break
-                elif number_of_peaks[idx] % 2 == 0:
-                    right_side_peak = number_of_peaks[idx]//2
[…]
