### 2020-09-28 f430b5c
Started to add CPU and GPU version of the same toolbox to one library. This allows to change between the CPU and GPU implementation on the fly and will allow the library to be installed on computers without CUDA. The CPU version will obviously be hell of a lot slower but can still be used.
Signed-off-by: Jan Andre Reuter <jan.andre.reuter@hotmail.de>
