+++ SLIX/toolbox.py
@@ -1,11 +1,9 @@
-import numpy
-import cupy
-from numba import cuda
-
-import nibabel
 import tifffile
-from SLIX._toolbox import _direction, _prominence, _peakwidth
+import nibabel
+import numpy
 
+from SLIX.SLIX_GPU import toolbox as gpu_toolbox
+from SLIX.SLIX_CPU import toolbox as cpu_toolbox
 
 def read_image(filepath):
     """
@@ -30,125 +28,22 @@ def read_image(filepath):
     return data
 
 
-def peaks(image, return_numpy=True):
-    gpu_image = cupy.array(image, dtype='float32')
-    right = cupy.roll(gpu_image, 1, axis=-1) - gpu_image
-    left = cupy.roll(gpu_image, -1, axis=-1) - gpu_image
-    del gpu_image
-
-    peaks = (left < 0) & (right <= 0)
-    del right
-    del left
-
-    if return_numpy:
-        peaks_cpu = cupy.asnumpy(peaks)
-        del peaks
-
-        return peaks_cpu
+def peaks(image, use_gpu=True, return_numpy=True):
+    if use_gpu:
+        return gpu_toolbox.peaks(image, return_numpy)
     else:
-        return peaks
+        return cpu_toolbox.peaks(image)
 
 
-def num_peaks(image, return_numpy=True):
-    gpu_image = cupy.array(image, dtype='float64')
-    right = cupy.roll(gpu_image, 1, axis=-1) - gpu_image
-    left = cupy.roll(gpu_image, -1, axis=-1) - gpu_image
-    del gpu_image
-
-    peaks = (left < 0) & (right <= 0)
-    del right
-    del left
-
-    resulting_image = cupy.empty((peaks.shape[:2]))
-    resulting_image[:, :] = cupy.count_nonzero(peaks, axis=-1)
-    if return_numpy:
-        resulting_image_cpu = cupy.asnumpy(resulting_image)
-        del resulting_image
-        return resulting_image_cpu
+def direction(peak_image, number_of_directions=3, use_gpu=True, return_numpy=True):
+    if use_gpu:
+        return gpu_toolbox.direction(peak_image, number_of_directions, return_numpy)
     else:
-        return resulting_image
+        return cpu_toolbox.direction(peak_image, number_of_directions)
 
 
-def normalize(image, kind_of_normalization=0, return_numpy=True):
-    gpu_image = cupy.array(image, dtype='float32')
-    if kind_of_normalization == 0:
-        gpu_image = (gpu_image - cupy.min(gpu_image, axis=-1)[..., None]) / \
-                    (cupy.max(gpu_image, axis=-1)[..., None] - cupy.min(gpu_image, axis=-1)[..., None])
+def peak_prominence(image, peak_image=None, kind_of_normalization=0, use_gpu=True, return_numpy=True):
+    if use_gpu:
+        return gpu_toolbox.peak_prominence(image, peak_image, kind_of_normalization, return_numpy)
     else:
-        gpu_image = gpu_image / cupy.mean(gpu_image, axis=-1)[..., None]
-
-    if return_numpy:
-        gpu_image_cpu = cupy.asnumpy(gpu_image)
-        del gpu_image
-        return gpu_image_cpu
-    else:
-        return gpu_image
-
-
-
-def peak_prominence(image, peak_image=None, kind_of_normalization=0, return_numpy=True):
-    gpu_image = cupy.array(image, dtype='float32')
-    if peak_image is not None:
-        gpu_peak_image = cupy.array(peak_image)
-    else:
-        gpu_peak_image = peaks(gpu_image, return_numpy=False)
-    gpu_image = normalize(gpu_image, kind_of_normalization, return_numpy=False)
-
[…]
