### 2020-09-28 7102a1f
Added peak width (full image and mean value)
Signed-off-by: Jan Andre Reuter <jan.andre.reuter@hotmail.de>
