### 2020-09-28 96ff39a
Added peak distance
Signed-off-by: Jan Andre Reuter <jan.andre.reuter@hotmail.de>
