### 2020-09-29 71d2198
Added CPU implementation of all GPU versions which are currently implemented. The calculation method can be changed with the USE_GPU option
