+++ SLIX/SLIX_CPU/_toolbox.py
@@ -1,0 +1,141 @@
+from numba import jit, prange
+import numpy
+
+# DEFAULT PARAMETERS
+BACKGROUND_COLOR = -1
+MAX_DISTANCE_FOR_CENTROID_ESTIMATION = 2
+
+NUMBER_OF_SAMPLES = 100
+TARGET_PEAK_HEIGHT = 0.94
+TARGET_PROMINENCE = 0.08
+
+
+@jit(nopython=True, parallel=True)
+def _prominence(image, peak_image):
+    result_image = numpy.empty(image.shape, dtype=numpy.float32)
+    for idx in prange(image.shape[0]):
+        sub_image = image[idx]
+        sub_peak_array = peak_image[idx]
+        for pos in prange(len(sub_peak_array)):
+            if sub_peak_array[pos] == 1:
+                i_min = -len(sub_peak_array) / 2
+                i_max = int(len(sub_peak_array) * 1.5)
+
+                i = pos
+                left_min = sub_image[pos]
+                while i_min <= i and sub_image[i] <= sub_image[pos]:
+                    if sub_image[i] < left_min:
+                        left_min = sub_image[i]
+                    i = i - 1
+
+                i = pos
+                right_min = sub_image[pos]
+                while i <= i_max and sub_image[i % len(sub_peak_array)] <= sub_image[pos]:
+                    if sub_image[i % len(sub_peak_array)] < right_min:
+                        right_min = sub_image[i % len(sub_peak_array)]
+                    i = i + 1
+
+                result_image[idx, pos] = sub_image[pos] - max(left_min, right_min)
+            else:
+                result_image[idx, pos] = 0
+    return result_image
+
+
+@jit(nopython=True, parallel=True, fastmath=True, nogil=True)
+def _peakwidth(image, peak_image, prominence, target_height):
+    result_image = numpy.empty(image.shape).astype(numpy.float32)
+    for idx in prange(image.shape[0]):
+        sub_image = image[idx]
+        sub_peak_array = peak_image[idx]
+        sub_prominece = prominence[idx]
+
+        for pos in prange(len(sub_peak_array)):
+            if sub_peak_array[pos] == 1:
+                height = sub_image[pos] - sub_prominece[pos] * target_height
+                i_min = -len(sub_peak_array) / 2
+                i_max = int(len(sub_peak_array) * 1.5)
+
+                i = int(pos)
+                while i_min < i and height < sub_image[i]:
+                    i -= 1
+                left_ip = float(i)
+                if sub_image[i] < height:
+                    # Interpolate if true intersection height is between samples
+                    left_ip += (height - sub_image[i]) / (sub_image[i + 1] - sub_image[i])
+
+                # Find intersection point on right side
+                i = int(pos)
+                while i < i_max and height < sub_image[i]:
+                    i += 1
+                right_ip = float(i)
+                if sub_image[i] < height:
+                    # Interpolate if true intersection height is between samples
+                    right_ip -= (height - sub_image[i]) / (sub_image[i - 1] - sub_image[i])
+
+                result_image[idx, pos] = right_ip - left_ip
+            else:
+                result_image[idx, pos] = 0
+    return result_image
+
+
+@jit(nopython=True, parallel=True)
+def _peakdistance(peak_image, number_of_peaks):
+    result_image = numpy.empty(peak_image.shape).astype(numpy.float32)
+    for idx in prange(peak_image.shape[0]):
+        sub_peak_array = peak_image[idx]
+        current_pair = 0
+
+        for i in prange(len(sub_peak_array)):
+            if sub_peak_array[i] == 1:
+                if number_of_peaks[idx] == 1:
+                    result_image[idx, i] = 360.0
+                    break
[…]
