+++ SLIX/toolbox.py
@@ -5,6 +5,7 @@ import numpy
 from SLIX.SLIX_GPU import toolbox as gpu_toolbox
 from SLIX.SLIX_CPU import toolbox as cpu_toolbox
 
+
 def read_image(filepath):
     """
     Reads image file and returns it.
@@ -42,11 +43,18 @@ def direction(peak_image, number_of_directions=3, use_gpu=True, return_numpy=Tru
         return cpu_toolbox.direction(peak_image, number_of_directions)
 
 
-def peakdistance(peak_image, use_gpu=True, return_numpy=True):
+def peak_distance(peak_image, use_gpu=True, return_numpy=True):
     if use_gpu:
-        return gpu_toolbox.peakdistance(peak_image, return_numpy)
+        return gpu_toolbox.peak_distance(peak_image, return_numpy)
     else:
-        return cpu_toolbox.peakdistance(peak_image)
+        return cpu_toolbox.peak_distance(peak_image)
+
+
+def mean_peak_distance(peak_image, use_gpu=True, return_numpy=True):
+    if use_gpu:
+        return gpu_toolbox.mean_peak_distance(peak_image, return_numpy)
+    else:
+        return cpu_toolbox.mean_peak_distance(peak_image)
 
 
 def peak_prominence(image, peak_image=None, kind_of_normalization=0, use_gpu=True, return_numpy=True):
@@ -56,8 +64,22 @@ def peak_prominence(image, peak_image=None, kind_of_normalization=0, use_gpu=Tru
         return cpu_toolbox.peak_prominence(image, peak_image, kind_of_normalization)
 
 
+def mean_peak_prominence(image, peak_image=None, kind_of_normalization=0, use_gpu=True, return_numpy=True):
+    if use_gpu:
+        return gpu_toolbox.mean_peak_prominence(image, peak_image, kind_of_normalization, return_numpy)
+    else:
+        return cpu_toolbox.mean_peak_prominence(image, peak_image, kind_of_normalization)
+
+
 def peak_width(image, peak_image=None, target_height=0.5, use_gpu=True, return_numpy=True):
     if use_gpu:
         return gpu_toolbox.peak_width(image, peak_image, target_height, return_numpy=return_numpy)
     else:
-        return gpu_toolbox.peak_width(image, peak_image, target_height)
+        return cpu_toolbox.peak_width(image, peak_image, target_height)
+
+
+def mean_peak_width(image, peak_image=None, target_height=0.5, use_gpu=True, return_numpy=True):
+    if use_gpu:
+        return gpu_toolbox.mean_peak_width(image, peak_image, target_height, return_numpy=return_numpy)
+    else:
+        return cpu_toolbox.mean_peak_width(image, peak_image, target_height)

[…]
