### 2020-09-29 7146b04
Added first part of centroid correction. Returning the image and the sampling part is missing though
Signed-off-by: Jan Andre Reuter <j.reuter@fz-juelich.de>
