### 2020-09-30 61d0972
Fixed some bugs in centorid calculation. Left bases and right bases are now introduced which will be used for the centroid calculation. Sampling is still missing.
Signed-off-by: Jan Andre Reuter <jan.andre.reuter@hotmail.de>
