### 2020-09-30 70f2a72
First full version of centroid correction implemented on the GPU side. CPU side is missing and will be implemented soon. Centroid correction needs to be verified with the SLIX package
Signed-off-by: Jan Andre Reuter <jan.andre.reuter@hotmail.de>
