### 2020-09-30 c58d198
Integrated centroid calculation into direction and peak distance. Difference to current SLIX direction is about 1-2 degrees most of the time. The centroid calculation will be corrected soon. Currently the centroid will be calculated using more points than needed
Signed-off-by: Jan Andre Reuter <jan.andre.reuter@hotmail.de>
