### 2020-10-01 8e2e2bf
Current calculation of features is closer to the original SLIX toolbox. Identified that the current peak calculation does not match the Scipy implementation.
An example:

Array is [0, 255, 255, 255, 0]
scipy.signal.find_peaks() returns [2]
current implementation will return [1]

Signed-off-by: Jan Andre Reuter <jan.andre.reuter@hotmail.de>
