### 2020-10-01 7f76659
Reworked main.py to support command line arguments. This is not fully finished yet.
Signed-off-by: Jan Andre Reuter <jan.andre.reuter@hotmail.de>
