### 2020-10-02 b06c1ae
Fixed wrong orientation of some images
Signed-off-by: Jan Andre Reuter <jan.andre.reuter@hotmail.de>
