### 2020-10-02 43690e1
First version of Scipy implementation of peak finding. Current version is still bugged
Signed-off-by: Jan Andre Reuter <jan.andre.reuter@hotmail.de>
