### 2020-10-02 99a14d2
Probably final peak calculation
Signed-off-by: Jan Andre Reuter <jan.andre.reuter@hotmail.de>
