+++ SLIX/SLIX_GPU/_toolbox.py
@@ -17,12 +17,13 @@ def _peak_cleanup(peaks, resulting_peaks):
     pos = 0
     while pos < len(sub_peak_array):
         if sub_peak_array[pos] == 1:
-            for offset in range(1, len(sub_peak_array)):
-                if sub_peak_array[(pos + offset) % len(sub_peak_array)] == 0:
-                    break
-                # sub_peak_array[(pos + offset) % len(sub_peak_array)] = 0
-            resulting_peaks[idx, (pos + (offset-1) // 2) % len(sub_peak_array)] = 1
-            pos = pos + offset
+            sub_peak_array[pos] = 0
+            offset = 1
+            while sub_peak_array[(pos + offset) % len(sub_peak_array)] == 1:
+                resulting_peaks[idx, (pos + offset) % len(sub_peak_array)] = 0
+                offset = offset + 1
+            resulting_peaks[idx, (pos + offset // 2) % len(sub_peak_array)] = 1
+            pos = pos + offset + 1
         else:
             resulting_peaks[idx, pos] = 0
             pos = pos + 1
