### 2020-10-02 d7ecc62
Corrected peak calculation
Signed-off-by: Jan Andre Reuter <jan.andre.reuter@hotmail.de>
