### 2020-10-05 cad3548
GPU version will now only be imported when the package CuPy is installed and a CUDA capable device is present
Signed-off-by: Jan Andre Reuter <jan.andre.reuter@hotmail.de>
