### 2020-10-05 5be1244
Removed no_gpu option. Added GPU/CPU selection based on hardware and installed packages. There might be a segmentation fault bug in there but it couldn't be replicated in multiple runthroughs after some changes.
Code still needs cleanup and testing. Also CPU vs GPU needs to be compared

Signed-off-by: Jan Andre Reuter <jan.andre.reuter@hotmail.de>
