+++ SLIX/SLIX_CPU/_toolbox.py
@@ -139,3 +139,85 @@ def _direction(peak_array, number_of_peaks, num_directions):
                     if current_direction == number_of_peaks[idx]//2:
                         break
     return result_image
+
+
+@jit(nopython=True, parallel=True)
+def _centroid_correction_bases(image, peak_image, reverse_peaks):
+    left_bases = numpy.empty(image.shape)
+    right_bases = numpy.empty(image.shape)
+
+    for idx in prange(image.shape[0]):
+        sub_image = image[idx]
+        sub_peaks = peak_image[idx]
+        sub_reverse_peaks = reverse_peaks[idx]
+
+        max_pos = 0
+        for pos in range(len(sub_image)):
+            if sub_image[pos] > max_pos:
+                max_pos = sub_image[pos]
+
+        for pos in range(len(sub_peaks)):
+            if sub_peaks[pos] == 1:
+
+                target_peak_height = max(0, sub_image[pos] - max_pos * (1 - TARGET_PEAK_HEIGHT))
+                left_position = MAX_DISTANCE_FOR_CENTROID_ESTIMATION
+                right_position = MAX_DISTANCE_FOR_CENTROID_ESTIMATION
+
+                # Check for minima in range
+                for offset in range(MAX_DISTANCE_FOR_CENTROID_ESTIMATION):
+                    if sub_reverse_peaks[pos - offset] == 1:
+                        left_position = offset
+                    if sub_reverse_peaks[(pos + offset) % len(sub_reverse_peaks)] == 1:
+                        right_position = offset
+
+                # Check for peak height
+                for offset in range(abs(left_position)):
+                    if sub_image[pos - offset] < target_peak_height:
+                        left_position = offset
+                        break
+                for offset in range(right_position):
+                    if sub_image[(pos + offset) % len(sub_image)] < target_peak_height:
+                        right_position = offset
+                        break
+
+                left_bases[idx, pos] = left_position
+                right_bases[idx, pos] = right_position
+            else:
+                left_bases[idx, pos] = 0
+                right_bases[idx, pos] = 0
+    return left_bases, right_bases
+
+
+@jit(nopython=True, parallel=True)
+def _centroid(image, peak_image, left_bases, right_bases):
+    centroid_peaks = numpy.empty(image.shape)
+
+    for idx in prange(image.shape[0]):
+        sub_image = image[idx]
+        sub_peaks = peak_image[idx]
+        sub_left_bases = left_bases[idx]
+        sub_right_bases = right_bases[idx]
+
+        for pos in range(len(sub_peaks)):
+            if sub_peaks[pos] == 1:
+                centroid_sum_top = 0.0
+                centroid_sum_bottom = 0.0
+                for x in range(-sub_left_bases[pos], sub_right_bases[pos]):
+                    img_pixel = sub_image[(pos + x) % len(sub_image)]
+                    next_img_pixel = sub_image[(pos + x + 1) % len(sub_image)]
+                    for interp in range(NUMBER_OF_SAMPLES+1):
+                        step = interp / NUMBER_OF_SAMPLES
+                        func_val = img_pixel + (next_img_pixel - img_pixel) * step
+                        if func_val > sub_peaks[pos] * TARGET_PEAK_HEIGHT:
+                            centroid_sum_top += (x + step) * func_val
+                            centroid_sum_bottom += func_val
+
+                centroid = centroid_sum_top / centroid_sum_bottom
+                if centroid > 1:
+                    centroid = 1
+                if centroid < -1:
+                    centroid = -1
+                centroid_peaks[idx, pos] = centroid
+            else:
+                centroid_peaks[idx, pos] = 0
+    return centroid_peaks
