+++ SLIX/toolbox.py
@@ -3,14 +3,14 @@ import nibabel
 import numpy
 
 try:
-    import cupy
-    cupy.empty(0)
-    from SLIX.SLIX_GPU import toolbox as gpu_toolbox
-except cupy.cuda.runtime.CUDARuntimeError:
-    pass
+    try:
+        import cupy
+        cupy.empty(0)
+        from SLIX.SLIX_GPU import toolbox as gpu_toolbox
+    except cupy.cuda.runtime.CUDARuntimeError:
+        pass
 except ModuleNotFoundError:
     pass
-
 from SLIX.SLIX_CPU import toolbox as cpu_toolbox
 
 
@@ -96,6 +96,6 @@ def mean_peak_width(image, peak_image=None, target_height=0.5, use_gpu=True, ret
 def centroid_correction(image, peak_image, low_prominence=cpu_toolbox.TARGET_PROMINENCE, high_prominence=None,
                         use_gpu=True, return_numpy=True):
     if use_gpu:
-        return gpu_toolbox.centroid_correction(image, peak_image, gpu_toolbox.TARGET_PROMINENCE, None, True)
+        return gpu_toolbox.centroid_correction(image, peak_image, low_prominence, high_prominence, return_numpy)
     else:
-        return cpu_toolbox.centroid_correction(image, peak_image, gpu_toolbox.TARGET_PROMINENCE, None)
+        return cpu_toolbox.centroid_correction(image, peak_image, low_prominence, high_prominence)

[…]
