### 2020-10-06 4585474
Implemented first tests. Removed bug in _toolbox for peak detection.
BEWARE: There is a bug in the GPU version for low amounts of line profiles where threads of the GPU can interfere one another. This needs to be fixed in a future version with a corresponding test!

Signed-off-by: Jan Andre Reuter <jan.andre.reuter@hotmail.de>
