### 2020-10-06 486cc05
Reimplemented parts of the test cases because thread size in toolbox.py caused problems. Simplified source code outside of Numba CUDA kernels.
TODO: Check thread sizes and choose number based on division through thread size.

Signed-off-by: Jan Andre Reuter <jan.andre.reuter@hotmail.de>
