+++ SLIX/SLIX_GPU/toolbox.py
@@ -1,5 +1,6 @@
 import cupy
 import numpy
+import math
 from numba import cuda
 
 from SLIX.SLIX_GPU._toolbox import _direction, _prominence, _peakwidth, _peakdistance, TARGET_PROMINENCE, \
@@ -7,34 +8,22 @@ from SLIX.SLIX_GPU._toolbox import _direction, _prominence, _peakwidth, _peakdis
 
 
 def peaks(image, return_numpy=True):
-    gpu_image = cupy.array(image, dtype='int32')
+    gpu_image = cupy.array(image, dtype='float32')
     right = cupy.roll(gpu_image, 1, axis=-1) - gpu_image
     left = cupy.roll(gpu_image, -1, axis=-1) - gpu_image
+    peaks = (left <= 0) & (right <= 0) & (gpu_image != 0)
     del gpu_image
 
-    peaks = (left <= 0) & (right <= 0)
-    print(peaks)
-    reshape = False
-    if len(peaks.shape) == 3:
-        reshape = True
-        [image_x, image_y, image_z] = peaks.shape
-        peaks = peaks.reshape(image_x * image_y, image_z)
-    else:
-        [image_x, image_z] = peaks.shape
-        image_y = 1
     del right
     del left
 
     resulting_peaks = cupy.zeros(peaks.shape, dtype='int8')
-    threads_per_block = 256
-    blocks_per_grid = (image_x * image_y + (threads_per_block - 1)) // threads_per_block
-    print(resulting_peaks.shape, peaks.shape)
+    threads_per_block = (1, 1)
+    blocks_per_grid = image.shape[:-1]
     _peak_cleanup[blocks_per_grid, threads_per_block](peaks, resulting_peaks)
-    print(resulting_peaks)
-
-    if reshape:
-        resulting_peaks = resulting_peaks.reshape(image_x, image_y, image_z)
+    cuda.synchronize()
     del peaks
+
 
     if return_numpy:
         peaks_cpu = cupy.asnumpy(resulting_peaks)
@@ -48,8 +37,7 @@ def peaks(image, return_numpy=True):
 def num_peaks(image, return_numpy=True):
     peak_image = peaks(image, return_numpy=False)
 
-    resulting_image = cupy.empty((peak_image.shape[:2]))
-    resulting_image[:, :] = cupy.count_nonzero(peak_image, axis=-1)
+    resulting_image = cupy.count_nonzero(peak_image, axis=-1)
     if return_numpy:
         resulting_image_cpu = cupy.asnumpy(resulting_image)
         del resulting_image
@@ -77,31 +65,18 @@ def normalize(image, kind_of_normalization=0, return_numpy=True):
 def peak_prominence(image, peak_image=None, kind_of_normalization=0, return_numpy=True):
     gpu_image = cupy.array(image, dtype='float32')
     if peak_image is not None:
-        gpu_peak_image = cupy.array(peak_image)
+        gpu_peak_image = cupy.array(peak_image).astype('uint8')
     else:
-        gpu_peak_image = peaks(gpu_image, return_numpy=False)
+        gpu_peak_image = peaks(gpu_image, return_numpy=False).astype('uint8')
     gpu_image = normalize(gpu_image, kind_of_normalization, return_numpy=False)
 
-    reshape = False
-    if len(image.shape) == 3:
-        reshape = True
-        [image_x, image_y, image_z] = image.shape
-    else:
-        [image_x, image_z] = image.shape
-        image_y = 1
-
-    gpu_image = gpu_image.reshape(image_x * image_y, image_z)
-    gpu_peak_image = gpu_peak_image.reshape(image_x * image_y, image_z).astype('int8')
-    result_img_gpu = cupy.empty((image_x * image_y, image_z), dtype='float32')
[…]
