+++ tests/test_toolbox_gpu.py
@@ -1,0 +1,216 @@
+from SLIX.SLIX_GPU import toolbox
+import numpy
+from matplotlib import pyplot as plt
+
+
+class TestToolbox:
+    def test_all_peaks(self):
+        # Create an absolute simple peak array
+        arr = numpy.array([0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0] * 260)
+        arr = arr.reshape((10, 26, 11))
+        real_peaks = arr == 1
+        toolbox_peaks = toolbox.peaks(arr)
+        assert numpy.all(toolbox_peaks == real_peaks)
+
+        # Test one single peak
+        arr = numpy.array(([0, 1, 0, 0, 0, 0]) * 260,
+                               dtype=bool)
+        arr = arr.reshape((10, 26, 6))
+
+        real_peaks = arr == 1
+        toolbox_peaks = toolbox.peaks(arr)
+        assert numpy.all(toolbox_peaks == real_peaks)
+
+
+    def test_num_peaks(self):
+        # Create an absolute simple peak array
+        test_arr = numpy.array(([0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0]) * 1000000, dtype=bool)
+        test_arr = test_arr.reshape((1000, 1000, 20))
+
+        real_peaks = test_arr == True
+        toolbox_peaks = toolbox.num_peaks(real_peaks)
+        expected_value = numpy.count_nonzero(real_peaks[0, 0, :])
+        assert numpy.all(toolbox_peaks == expected_value)
+
+
+    def test_peak_prominence(self):
+        # Create an absolute simple peak array
+        arr = numpy.array([0, 1, 0, 0.07, 0, 1, 0, 0.07, 0, 1, 0] * 1000000)
+        arr = arr.reshape((1000, 1000, 11))
+        # Test if high and low prominence separation is working as intended
+        high_peaks = arr == 1
+        low_peaks = arr == 0.07
+
+        toolbox_peaks = toolbox.peaks(arr)
+        toolbox_prominence = toolbox.peak_prominence(arr, toolbox_peaks)
+        toolbox_high_peaks = toolbox_peaks.copy()
+        toolbox_high_peaks[toolbox_prominence < toolbox.TARGET_PROMINENCE] = False
+        toolbox_low_peaks = toolbox_peaks.copy()
+        toolbox_low_peaks[toolbox_prominence >= toolbox.TARGET_PROMINENCE] = False
+        assert numpy.all(high_peaks == toolbox_high_peaks)
+        assert numpy.all(low_peaks == toolbox_low_peaks)
+
+    """def test_peakdistance(self):
+        # Test one peak
+        test_arr = numpy.array(([True, False, False] + [False] * 21) * 1000000)
+        test_arr = test_arr.reshape((1000, 1000, 24))
+        expected_distance = 360
+
+        toolbox_peaks = toolbox.peaks(test_arr)
+        toolbox_distance = toolbox.peak_distance(toolbox_peaks, numpy.zeros(toolbox_peaks.shape, dtype=float))
+        toolbox_distance_reduced = numpy.sum(toolbox_distance, axis=-1)
+        assert numpy.all(toolbox_distance_reduced == expected_distance)
+
+        test_arr = numpy.array(([False, False, True, False, False, False, False, True, False] + [False] * 15) * 1000000)
+        test_arr = test_arr.reshape((1000, 1000, 24))
+        expected_distance = 75
+
+        toolbox_peaks = toolbox.peaks(test_arr)
+        toolbox_distance = toolbox.peak_distance(toolbox_peaks, numpy.zeros(toolbox_peaks.shape, dtype=float))
+        assert numpy.count_nonzero(toolbox_distance)
[…]

[…]
