### 2020-10-07 21aa2e8
Added peakdistance test
Signed-off-by: Jan Andre Reuter <jan.andre.reuter@hotmail.de>
