### 2020-10-07 46601bd
Added peakwidth test. Reduced test load while maintaining same test
Signed-off-by: Jan Andre Reuter <jan.andre.reuter@hotmail.de>
