+++ tests/test_toolbox_gpu.py
@@ -6,25 +6,23 @@ from matplotlib import pyplot as plt
 class TestToolbox:
     def test_all_peaks(self):
         # Create an absolute simple peak array
-        arr = numpy.array([0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0] * 260)
-        arr = arr.reshape((10, 26, 11))
+        arr = numpy.array([0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0] * 1)
+        arr = arr.reshape((1, 1, 11))
         real_peaks = arr == 1
         toolbox_peaks = toolbox.peaks(arr)
         assert numpy.all(toolbox_peaks == real_peaks)
 
         # Test one single peak
-        arr = numpy.array(([0, 1, 0, 0, 0, 0]) * 260,
-                               dtype=bool)
-        arr = arr.reshape((10, 26, 6))
-
+        arr = numpy.array(([0, 1, 0, 0, 0, 0]), dtype=bool)
+        arr = arr.reshape((1, 1, 6))
         real_peaks = arr == 1
         toolbox_peaks = toolbox.peaks(arr)
         assert numpy.all(toolbox_peaks == real_peaks)
 
     def test_num_peaks(self):
         # Create an absolute simple peak array
-        test_arr = numpy.array(([0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0]) * 1000000, dtype=bool)
-        test_arr = test_arr.reshape((1000, 1000, 20))
+        test_arr = numpy.array(([0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0]), dtype=bool)
+        test_arr = test_arr.reshape((1, 1, 20))
 
         real_peaks = test_arr == True
         toolbox_peaks = toolbox.num_peaks(real_peaks)
@@ -33,8 +31,8 @@ class TestToolbox:
 
     def test_peak_prominence(self):
         # Create an absolute simple peak array
-        arr = numpy.array([0, 1, 0, 0.07, 0, 1, 0, 0.07, 0, 1, 0] * 1000000)
-        arr = arr.reshape((1000, 1000, 11))
+        arr = numpy.array([0, 1, 0, 0.07, 0, 1, 0, 0.07, 0, 1, 0] * 1)
+        arr = arr.reshape((1, 1, 11))
         # Test if high and low prominence separation is working as intended
         high_peaks = arr == 1
         low_peaks = arr == 0.07
@@ -50,8 +48,8 @@ class TestToolbox:
 
     def test_peakdistance(self):
         # Test one peak
-        test_arr = numpy.array(([True, False, False] + [False] * 21) * 1000000)
-        test_arr = test_arr.reshape((1000, 1000, 24))
+        test_arr = numpy.array(([True, False, False] + [False] * 21))
+        test_arr = test_arr.reshape((1, 1, 24))
         expected_distance = 360
 
         toolbox_peaks = toolbox.peaks(test_arr)
@@ -60,42 +58,41 @@ class TestToolbox:
         assert numpy.all(toolbox_distance_reduced == expected_distance)
 
         # Test two peaks
-        test_arr = numpy.array(([False, False, True, False, False, False, False, True, False] + [False] * 15) * 1000000)
-        test_arr = test_arr.reshape((1000, 1000, 24))
+        test_arr = numpy.array(([False, False, True, False, False, False, False, True, False] + [False] * 15))
+        test_arr = test_arr.reshape((1, 1, 24))
         expected_distance = 75
 
[…]
