### 2020-10-07 fd35567
Fixed a bug where the GPU direction calculation could fail because of a wrong dtype. Added direction test.
Signed-off-by: Jan Andre Reuter <jan.andre.reuter@hotmail.de>
