### 2020-10-07 0664702
Added normalize
Signed-off-by: Jan Andre Reuter <jan.andre.reuter@hotmail.de>
