### 2020-10-07 b237ba6
Added new test files. Added centroid calculation tests. Changed maximum width for centroid calculation to 3 as wide peaks are estimated better with it.
Signed-off-by: Jan Andre Reuter <jan.andre.reuter@hotmail.de>
