### 2020-10-07 242d517
Readded some of the CPU tests. CPU version still needs to be fixed.
Readded requirements.txt

Signed-off-by: Jan Andre Reuter <jan.andre.reuter@hotmail.de>
