+++ SLIX/SLIX_CPU/_toolbox.py
@@ -3,7 +3,7 @@ import numpy
 
 # DEFAULT PARAMETERS
 BACKGROUND_COLOR = -1
-MAX_DISTANCE_FOR_CENTROID_ESTIMATION = 2
+MAX_DISTANCE_FOR_CENTROID_ESTIMATION = 3
 
 NUMBER_OF_SAMPLES = 100
 TARGET_PEAK_HEIGHT = 0.94
@@ -23,17 +23,21 @@ def _prominence(image, peak_image):
 
                 i = pos
                 left_min = sub_image[pos]
-                while i_min <= i and sub_image[i] <= sub_image[pos]:
+                wlen = len(sub_peak_array) - 1
+                while i_min <= i and sub_image[i] <= sub_image[pos] and wlen > 0:
                     if sub_image[i] < left_min:
                         left_min = sub_image[i]
                     i = i - 1
+                    wlen = wlen - 1
 
                 i = pos
                 right_min = sub_image[pos]
-                while i <= i_max and sub_image[i % len(sub_peak_array)] <= sub_image[pos]:
+                wlen = len(sub_peak_array) - 1
+                while i <= i_max and sub_image[i % len(sub_peak_array)] <= sub_image[pos] and wlen > 0:
                     if sub_image[i % len(sub_peak_array)] < right_min:
                         right_min = sub_image[i % len(sub_peak_array)]
                     i = i + 1
+                    wlen = wlen - 1
 
                 result_image[idx, pos] = sub_image[pos] - max(left_min, right_min)
             else:
@@ -79,7 +83,7 @@ def _peakwidth(image, peak_image, prominence, target_height):
 
 
 @jit(nopython=True, parallel=True)
-def _peakdistance(peak_image, number_of_peaks):
+def _peakdistance(peak_image, centroids, number_of_peaks):
     result_image = numpy.empty(peak_image.shape).astype(numpy.float32)
     for idx in prange(peak_image.shape[0]):
         sub_peak_array = peak_image[idx]
@@ -110,7 +114,7 @@ def _peakdistance(peak_image, number_of_peaks):
 
 
 @jit(nopython=True, parallel=True)
-def _direction(peak_array, number_of_peaks, num_directions):
+def _direction(peak_array, centroids, number_of_peaks, num_directions):
     result_image = numpy.empty((peak_array.shape[0], num_directions)).astype(numpy.float32)
     for idx in prange(peak_array.shape[0]):
         sub_peak_array = peak_array[idx]
