+++ .gitlab-ci.yml
@@ -1,6 +1,7 @@
 image: "python:3.7"
 
 before_script:
+    - lspci | grep VGA
     - python --version
     - pip install pylint flake8
     - pip install -r requirements.txt
