### 2020-10-07 b03a522
Update .gitlab-ci.yml
