+++ .gitlab-ci.yml
@@ -1,6 +1,9 @@
-image: "python:3.7"
+image: "ubuntu:latest"
 
 before_script:
+    - export DEBIAN_FRONTEND=noninteractive
+    - apt update && apt-get upgrade -y
+    - apt install python3 python3-pip
     - lspci | grep VGA
     - python --version
     - pip install pylint flake8
