+++ .gitlab-ci.yml
@@ -2,12 +2,12 @@ image: "ubuntu:latest"
 
 before_script:
     - export DEBIAN_FRONTEND=noninteractive
-    - apt update && apt-get upgrade -y
-    - apt install python3 python3-pip
+    - apt-get update -qq && apt-get upgrade -y
+    - apt-get install -y python3 python3-pip
     - lspci | grep VGA
-    - python --version
-    - pip install pylint flake8
-    - pip install -r requirements.txt
+    - python3 --version
+    - pip3 install pylint flake8
+    - pip3 install -r requirements.txt
 
 stages:
     - Analysis
