+++ .gitlab-ci.yml
@@ -3,7 +3,7 @@ image: "ubuntu:latest"
 before_script:
     - export DEBIAN_FRONTEND=noninteractive
     - apt-get update -qq && apt-get upgrade -y
-    - apt-get install -y python3 python3-pip
+    - apt-get install -y python3 python3-pip pciutils nvidia-cuda-toolkit
     - lspci | grep VGA
     - python3 --version
     - pip3 install pylint flake8
