### 2020-10-08 ec34d18
Fixed CPU tests.
Signed-off-by: Jan Andre Reuter <j.reuter@fz-juelich.de>
