### 2020-10-09 0fccf11
Removed __init__.py as it caused problems with tests and importing library without using the syntax `from ... import ...`
Added demo file for tests. Added first test

Signed-off-by: Jan Andre Reuter <jan.andre.reuter@hotmail.de>
