### 2020-10-09 2ce5e11
Expanded tests for peak detection.
Fixed some bugs regarding the CPU toolbox peak detection.
toolbox.py in each package can now be referenced via `import SLIX[...]; SLIX[...].toolbox.[...]`
Current version shows errors for centroid calculation and other metrics when comparing CPU and GPU version

Signed-off-by: Jan Andre Reuter <jan.andre.reuter@hotmail.de>
