### 2020-10-09 f92fdce
Fixed centroid calculation for CPU
Signed-off-by: Jan Andre Reuter <jan.andre.reuter@hotmail.de>
