### 2020-10-09 40f0306
Fixed peak width method for both versions.
Test will now pass.

Signed-off-by: Jan Andre Reuter <jan.andre.reuter@hotmail.de>
