### 2020-10-10 3d-pli/SLIX#1 @alexrockhill
[BUG, DOC] Running an example end of file error
I downloaded `Vervet1818_s0512_60um_SLI_090_Stack_1day.nii` from the link in the README example and tried to run it directly as input. In the example gif, I saw that the input took at `stack.tif`. Does `SLIXParameterGenerator` take a `nii` file as input, and where is it documented what input types are accepted? Also, it would be helpful to have clear instructions on how to run an example with this or another file.

If `SLIXParameterGenerator` does take an `nii` file as input (it did look like the command made it some ways), it hung up after completing the ROI command so I stopped execution. Any idea what may be going wrong?

```
(venv) Alexs-MacBook-Pro:bin alexrockhill$ ./SLIXParameterGenerator -i ../Vervet1818_s0512_60um_SLI_090_Stack_1day.nii -o ../output/
SLI Feature Generator:
Number of threads: 4

Chosen feature maps:
Direction maps: True
Peak maps: True
Peak prominence map: True
Peak width map: True
Peak distance map: True
Optional maps: False

../Vervet1818_s0512_60um_SLI_090_Stack_1day.nii
Roi finished

^CTraceback (most recent call last):
  File "./SLIXParameterGenerator", line 233, in generate_feature_maps
    resulting_parameter_maps[i, current_index] = toolbox.prominence(peak_positions_high_non_centroid, roi)
  File "/Users/alexrockhill/software/SLIX/venv/lib/python3.7/site-packages/SLIX-1.1-py3.7.egg/SLIX/toolbox.py", line 187, in prominence
  File "/Users/alexrockhill/software/SLIX/venv/lib/python3.7/site-packages/SLIX-1.1-py3.7.egg/SLIX/toolbox.py", line 657, in normalize
  File "<__array_function__ internals>", line 6, in mean
  File "/Users/alexrockhill/software/SLIX/venv/lib/python3.7/site-packages/numpy/core/fromnumeric.py", line 3372, in mean
    return _methods._mean(a, axis=axis, dtype=dtype,
KeyboardInterrupt

Traceback (most recent call last):
  File "./SLIXParameterGenerator", line 366, in <module>
    args['with_smoothing'], args['mask_threshold'])
  File "./SLIXParameterGenerator", line 69, in full_pipeline
    parameter_maps = generate_feature_maps(roiset, selected_methods)
  File "./SLIXParameterGenerator", line 249, in generate_feature_maps
    current_index += 3
  File "/Users/alexrockhill/software/SLIX/venv/lib/python3.7/site-packages/pymp/__init__.py", line 122, in __exit__
    self._exception_queue.put((exc_t, exc_val, self._thread_num))
  File "<string>", line 2, in put
  File "/Users/alexrockhill/software/anaconda3/envs/swannlab/lib/python3.7/multiprocessing/managers.py", line 819, in _callmethod
    kind, result = conn.recv()
  File "/Users/alexrockhill/software/anaconda3/envs/swannlab/lib/python3.7/multiprocessing/connection.py", line 250, in recv
    buf = self._recv_bytes()
  File "/Users/alexrockhill/software/anaconda3/envs/swannlab/lib/python3.7/multiprocessing/connection.py", line 407, in _recv_bytes
    buf = self._recv(4)
  File "/Users/alexrockhill/software/anaconda3/envs/swannlab/lib/python3.7/multiprocessing/connection.py", line 383, in _recv
    raise EOFError
EOFError
```
https://github.com/openjournals/joss-reviews/issues/2675
