### 2020-10-10 3d-pli/SLIX#1 @alexrockhill
I waited for an hour before interrupting execution so it's possible that the command may not have finished, if that is the case, it would be helpful to print to the console the progress of the function.
