### 2020-10-11 3d-pli/SLIX#1 @Thyre
Hi and thanks for your issue,

the toolbox does indeed accept NIfTI images. The .gif in the README does not represent this but it is noted under the 'Required parameters' section of 'SLIXParameterGenerator'. However, I can change the .gif to also include an example with a .nii file.
In it's current implementation, SLIX will take some time when calculating the parameter maps for full scale images. I can see that your MacBook has four cores and four threads. On my machine the program took around 10-15 minutes to generate all parameter maps. However, my desktop PC has 16 threads. Therefore I can imagine that the execution on your machine will take around 1 hour to 1.5 hours. I'm currently working on improving the performance by using CuPy and Numba (or solely Numba if an NVIDIA GPU isn't available), but can't guarantee a date for a new version yet.

In the meantime, I would recommended to run the evaluation on downsampled images, e.g. with --roisize 10, as suggested in the 'Resulting parameter maps' section. This means in your case:
./SLIXParameterGenerator -i ../Vervet1818_s0512_60um_SLI_090_Stack_1day.nii -o ../output/ --roisize 10
On my PC, this took only a few seconds.

In any case, it is a good idea to print to the console the progress of the function. I will look into that.
