### 2020-10-11 3d-pli/SLIX#1 @alexrockhill
Re speeding up the process, that would be excellent, but just from a user standpoint, I think there is an issue when the process takes a long time on a single core with a standard processor and there is no terminal output. It could also help with planning to describe how long the it might take on one or two example setups such a s how freesurfer does https://surfer.nmr.mgh.harvard.edu/fswiki/recon-all.
