### 2020-10-12 c5439c9
Delete .gitlab-ci.yml
