+++ .gitlab-ci.yml
@@ -1,32 +1,0 @@
-image: "ubuntu:latest"
-
-before_script:
-    - export DEBIAN_FRONTEND=noninteractive
-    - apt-get update -qq && apt-get upgrade -y
-    - apt-get install -y python3 python3-pip pciutils nvidia-cuda-toolkit
-    - lspci | grep VGA
-    - python3 --version
-    - pip3 install pylint flake8
-    - pip3 install -r requirements.txt
-
-stages:
-    - Analysis
-    - Test
-
-flake8:
-    stage: Analysis
-    allow_failure: true
-    script:
-        - flake8 --max-line-length=120 bin/* SLIX/*.py
-
-pylint:
-  stage: Analysis
-  allow_failure: true
-  script:
-  - pylint -d C0301 bin/* SLIX/*.py
-
-pytest:
-    stage: Test
-    script:
-        - pytest
-
