### 2020-10-12 088a6aa
Changed version number to v1.1.1
Signed-off-by: Jan Andre Reuter <jan.andre.reuter@hotmail.de>
