### 2020-10-12 f4b002b
Fixed a small bug within the sampling method.
The method create_sampling in SLIX/toolbox.py wasn't using the NUMBER_OF_SAMPLES parameter. Instead a fixed number was used. This is fixed now

Signed-off-by: Jan Andre Reuter <jan.andre.reuter@hotmail.de>
