### 2020-10-12 3d-pli/SLIX#1 @Thyre
I have added a progress bar to SLIX. It should give you an indication of how long the program takes to complete the input file. The README.md also has a new section with performance metrics of my PC. I might also add some runtimes of other computers, but I don't have access to another computer at the moment.
