### 2020-10-12 3d-pli/SLIX#1 @alexrockhill
Great, that would be very helpful to a first-time user running the example.

I think also what got lost in this being multiple issues in one is that, although I like the example gif, the user doesn't have that file and then when they grab the 100 MB nii that seems most obvious, it's not immediately apparent to downsample so I would suggest changing the example to a file the user has access to and giving the suggested command explicitly.
