### 2020-10-12 3d-pli/SLIX#1 @Thyre
I've changed the GIF to show examples of two measurements available in the [EBRAINS data repository](https://kg.ebrains.eu/search/instances/Dataset/de08dcb8-28d4-4f8f-b019-ef0f8924d5d4) and added the `roisize` option to make sure that the examples have a low execution time on most machines (max. one minute).
In addition to the GIF I added the needed commands to download both measurements and to generate the parameter maps as shown in the GIF.

I hope that this will make it clearer how to use the program and that downsampling is used for these examples.
