### 2020-10-12 3d-pli/SLIX#4 @alexrockhill
[DOC] Main Functionality Example
I ran the first example and got the following figures [output.zip](https://github.com/3d-pli/SLIX/files/5368001/output.zip). They look a bit sparse.

To make "a detailed reconstruction of the brain's nerve fiber architecture" as said in the paper, seems to me to be more like Figure 2i which I don't see in the output.

It would be helpful to have
1) explanations like those in the last section of the README for the examples and
2) the example generate a figure like 2i which I think would better qualify as a reconstruction of nerve architecture

Also, it is not immediately obvious how one would proceed with Figure 2i as input to do some kind of analysis. To bring up freesurfer as an example again, in the reconstruction you have different brain regions segmented out in volumetric boundaries. This seems like a fairly crucial part of a reconstruction; to have some format separating each nerve fiber. Perhaps I am not understanding the directional .tiff images appropriately if they in fact do just that.

https://github.com/openjournals/joss-reviews/issues/2675
