### 2020-10-12 3d-pli/SLIX#5 @alexrockhill
[DOC] Contributor Guidelines
According to JOSS policy, contribution guidelines are requested to be written.

https://github.com/openjournals/joss-reviews/issues/2675
