### 2020-10-12 3d-pli/SLIX#6 @alexrockhill
[DOC] Methods explanations
JOSS policy requires "A summary describing the high-level functionality and purpose of the software for a diverse, non-specialist audience."

Methodological choices such as why 8% prominence is used to qualify as a prominent peak or even why using prominent peaks is the best method for identifying crossing fibers don't seem readily understandable from the paper and documentation for a diverse audience.

Perhaps the best way to explain the methods would be to have an example or documentation walking the user through the process from raw data through each of the figures and to the output, considering parameter and processing choices along the way.

How and when to use `--with_mask` and `--with_smoothing` especially are not readily apparent.

EDIT: from reading appendix A of the Bioarvix paper, I see why you chose 8% prominence but it seems like the intensity range (Imax - Imin) may vary across setups, I'm not sure I would use 8% as a hard rule but would want to maybe check on some small section of the data that could be inspected manually or some other solution. Or maybe the goodness-of-fit and this and other parameters can be ascertained from Figure 8 (the parameter maps) but it is not clear from the documentation how to do so.

https://github.com/openjournals/joss-reviews/issues/2675
