### 2020-10-13 3d-pli/SLIX#7 @alexrockhill
[DOC] Integration with Fiber Tractography Software
In your bioarxiv paper, it says "We developed the open-source software SLIX (Scattered Light Imaging
ToolboX) that allows an automated evaluation of the measurement, and the computation of different parameter maps
containing various tissue information. The resulting fiber direction maps can be used, for example, as input for fiber
tractography algorithms."

From the perspective of a user, I'm not sure if SLIX meets the JOSS requirement that "that software should be feature-complete" without documentation and an example on how to integrate with a tractography software or algorithm. It looks like it is very close to being able to be used in a TRACULA pipeline for instance (there might be a step or two missing such as coregistration), but it is not apparent, to me at least, how the SLIX output images would be integrated.

https://github.com/openjournals/joss-reviews/issues/2675
