### 2020-10-13 3d-pli/SLIX#8 @alexrockhill
[DOC] Understanding sample data
I think it's a bit hard to figure out how to use the code without explaining the sample data better. Are the images 24 light angles that are 1314 x 1176 pixels?

I assume you would get images in some image format from the microscope, but then it's not clear how you would convert that to nii with the light angle stored in the nii metadata. And for the .tiff files, how is the light angle stored? That it if indeed the third dimension is light angle.

I think maybe this should be included to explain how to get from raw data capture to results/parameter maps that can be used in many ways to find results.

```
import nibabel as nib
img = nib.load('SLI-human-Sub-01_2xOpticTracts_s0037_30um_SLI_105_Stack_3days_registered.nii')
img.shape
```
```
(1314, 1176, 24, 1)
```
https://github.com/openjournals/joss-reviews/issues/2675
