### 2020-10-13 b245b4e
Added first version of CONTRIBUTING.md
Signed-off-by: Jan Andre Reuter <jan.andre.reuter@hotmail.de>
