### 2020-10-13 84dd814
Added first version of CONTRIBUTING.md
Signed-off-by: Jan Andre Reuter <jan.andre.reuter@hotmail.de>
