### 2020-10-13 3d-pli/SLIX#5 @Thyre
I've added a first version of the contribution guidelines [here](https://github.com/3d-pli/SLIX/blob/master/CONTRIBUTING.md)
Is there a section missing in the document or is it okay?
