### 2020-10-13 566a9db
Update README.md
(include screenshots, show/explain all generated parameter maps)
