+++ README.md
@@ -66,7 +66,7 @@ Additional parameters that determine which parameter maps will be generated from
 | `--optional`      | Generate four additional parameter maps: average value of each SLI profile (`_avg.tiff`), maximum value of each SLI profile (`_max.tiff`), minimum value of each SLI profile (`_min.tiff`), in-plane direction angles in regions without crossings (`_dir.tiff`). |
 
 ### Example
-The following example shows how to generate parameter maps from an SLI image stack, for two artificially crossing sections of human optic tracts (left) and a coronal section (upper left corner) of a vervet monkey brain (right): 
+The following example demonstrates the generation of parameter maps, for two artificially crossing sections of human optic tracts (left) and the upper left corner of a coronal section from a vervet monkey brain (right): 
 
 <img src="https://jugit.fz-juelich.de/j.reuter/slix/-/raw/assets/Screenshot_Demo1.png" height="327"><img src="https://jugit.fz-juelich.de/j.reuter/slix/-/raw/assets/Screenshot_Demo2.png" height="327">
 
@@ -94,7 +94,7 @@ SLIXParameterGenerator -i ./SLI-human-Sub-01_2xOpticTracts_s0037_30um_SLI_105_St
 SLIXParameterGenerator -i ./Vervet1818_s0512_60um_SLI_090_Stack_1day.nii -o . --roisize 10 --direction
 ```
 
-The execution of both commands should take around one minute max. The resulting parameter maps will be downsampled. For full resolution parameter maps do not use the `roisize` option. In this case, the computing time will be higher (around 25 times higher for the first example and 100 times higher for the second example).
+The execution of both commands should take around one minute max. The resulting parameter maps will be downsampled. To obtain full resolution parameter maps, do not use the `roisize` option. In this case, the computing time will be higher (around 25 times higher for the first example and 100 times higher for the second example).
 
 ### Resulting parameter maps
 
@@ -110,7 +110,7 @@ The parameter maps show in sequence: the average intensity of the SLI profiles (
  
[…]
