### 2020-10-13 022b193
Update README.md
(add more details for performance metrics)
