### 2020-10-13 4c65997
Update README.md
(minor changes)
