+++ README.md
@@ -48,10 +48,10 @@ SLIXParameterGenerator -i [INPUT-STACK] -o [OUTPUT-FOLDER] [[parameters]]
 | Parameter          | Function                                                                                                                                            |
 | ------------------ | --------------------------------------------------------------------------------------------------------------------------------------------------- |
 | `-r, --roisize`    | Average every NxN pixels of the SLI images and run the evaluation on the resulting (downsampled) images. Later on, the images will be upscaled to match the input file dimensions. (Default: N=1, i.e.`-r 1`) |
-| `--with_mask`      | Remove the background based on the maximum value of each image pixel. May include gray matter.                                                                |
-| `--mask_threshold` | Set the threshold for the background mask. Pixels for which the maximum intensity value of the SLI profile is below the threshold, will be considered as background. Higher values might remove the background better but will also include more of the gray matter. (Default = 10) |
+| `--with_mask`      | Remove the background (regions without tissue) based on the maximum value of each image pixel: Pixels for which the maximum intensity value of the SLI profile is below the threshold (defined by `--mask_threshold`), will be considered as background.                                                                |
+| `--mask_threshold` | Set the threshold for the background mask (`--with_mask`). Higher values might remove the background better but will also include more regions with gray matter. (Default = 10) |
 | `--num_procs`      | Run the program with the selected number of processes. (Default = either 16 threads or the maximum number of threads available.)                                  |
-| `--with_smoothing` | Apply smoothing to the SLI profiles for each image pixel before evaluation. The smoothing is performed using a Savitzky-Golay filter with 45 sampling points and a second order polynomial. (Designed for measurements with <img src="https://render.githubusercontent.com/render/math?math=\Delta\phi"> = 5° steps.)                                                                                     |
+| `--with_smoothing` | Apply smoothing to the SLI profiles for each image pixel before evaluation. The smoothing is performed using a Savitzky-Golay filter with 45 sampling points and a second order polynomial. (Designed for measurements with <img src="https://render.githubusercontent.com/render/math?math=\Delta\phi"> < 5° steps.)                                                                                     |
 
 ### Output
 Additional parameters that determine which parameter maps will be generated from the SLI image stack. If no parameter is used, the following parameter maps will be generated: peakprominence, number of peaks, peakwidth, peakdistance, direction angles in crossing regions. If `--optional` is used, four additional parameter maps will be generated (average, maximum, minimum, direction angles in non-crossing regions). If any parameter (except `–-optional`) is used, no parameter map besides the ones specified will be generated.
@@ -60,13 +60,13 @@ Additional parameters that determine which parameter maps will be generated from
 | -------------- | --------------------------------------------------------------------------- |
 | `--peakprominence`| Generate a parameter map (`_peakprominence.tiff`) containing the average prominence ([scipy.signal.peak_prominence](https://docs.scipy.org/doc/scipy/reference/generated/scipy.signal.peak_prominences.html#scipy.signal.peak_prominences)) of an SLI profile (image pixel), normalized by the average of the SLI profile. |
[…]
