+++ README.md
@@ -48,8 +48,8 @@ SLIXParameterGenerator -i [INPUT-STACK] -o [OUTPUT-FOLDER] [[parameters]]
 | Parameter          | Function                                                                                                                                            |
 | ------------------ | --------------------------------------------------------------------------------------------------------------------------------------------------- |
 | `-r, --roisize`    | Average every NxN pixels of the SLI images and run the evaluation on the resulting (downsampled) images. Later on, the images will be upscaled to match the input file dimensions. (Default: N=1, i.e.`-r 1`) |
-| `--with_mask`      | Remove the background (regions without tissue) based on the maximum value of each image pixel: Pixels for which the maximum intensity value of the SLI profile is below the threshold (defined by `--mask_threshold`), will be considered as background.                                                                |
-| `--mask_threshold` | Set the threshold for the background mask (`--with_mask`). Higher values might remove the background better but will also include more regions with gray matter. (Default = 10) |
+| `--with_mask`      | Remove the background (regions without tissue) based on the maximum value of each image pixel: Pixels for which the maximum intensity value of the SLI profile is below the threshold (defined by `--mask_threshold`), will be considered as background, i.e. the corresponding regions are set to zero for all images in the stack and will not be considered in the generation of the parameter maps.                                                                |
+| `--mask_threshold` | Set the threshold for the background mask (can only be used together with `--with_mask`). Higher values might remove the background better but will also include more regions with gray matter. (Default = 10) |
 | `--num_procs`      | Run the program with the selected number of processes. (Default = either 16 threads or the maximum number of threads available.)                                  |
 | `--with_smoothing` | Apply smoothing to the SLI profiles for each image pixel before evaluation. The smoothing is performed using a Savitzky-Golay filter with 45 sampling points and a second order polynomial. (Designed for measurements with <img src="https://render.githubusercontent.com/render/math?math=\Delta\phi"> < 5° steps.)                                                                                     |
 
