### 2020-10-13 3d-pli/SLIX#5 @alexrockhill
That seems to me to check all the boxes, thanks
