### 2020-10-14 3d-pli/SLIX#4 @miriammenzel
Thank you very much for your comment.

When looking at the parameter maps of the first example, you should keep in mind that the maps were generated with lower resolution (5 x 5 pixels are combined into one). Therefore, the parameter maps show less detail. You can generate the high-resolution maps by running the example without --roisize 5 (this takes 25 times longer).
In the “sparse” images, regions that cannot be evaluated are set to -1 (shown in white in Figure 2f,g,h): As described in the “Output” section of the README, the peakdistance (in “_peakdistance.tiff”) is only defined if the SLI profile has not exactly two prominent peaks; the in-plane direction angles (in “_dir,1,2,3.tiff”) are only defined if the SLI profile has two, four or six prominent peaks with a pair-wise distance of (180 +/- 35)°. The parameter map “_dir_1.tiff” contains the smallest detected fiber direction angle. If there exists only one fiber, its orientation is shown in “_dir_1.tiff”. The parameter map “_dir_2.tiff” (“_dir_3.tiff”) shows the second (third) fiber direction in regions with two (three) crossing fibers. As the sample does not contain many regions with two crossing fibers and no regions with three crossing fibers (the sample consists of two fiber bundles/optic tracts, artificially placed on top of each other), there are a lot of values with -1 within the tissue. For visualization purposes, Figure 2h in the JOSS paper shows the direction angle phi_1 in all image pixels where phi_2 is not defined (non-crossing regions).
We will point this out in the caption of Figure 2 and in the description of the parameter maps in the README. In addition to that, we will provide the full set of parameter maps and a more detailed description in the “Resulting parameter maps” section.

Concerning Figure 2i, this is only a visual representation of the data contained in the direction .tiff images (g,h): The direction angles for a certain region (here: 24 x 24 pixels) were represented by a line with the respective polar/direction angle. To generate the image, we used a simple FIJI Macro. In principle, any visualization tool that is able to represent multiple unit vectors as lines can be used. The visualization is not part of the SLIX software. If you like, we can provide the FIJI Macro to give users a more intuitive/visual understanding of the data?

The parameter maps generated by SLIX, especially the direction images, enhance the current reconstruction of nerve fiber orientations (obtained e.g. with 3D-Polarized Light Imaging), by adding crucial information about the orientations of crossing nerve fibers (in microscopic image pixels). Note that “reconstruction” is here used to describe the reconstruction of fiber orientations, i.e. a vector map containing different (crossing) nerve fiber orientations, not the nerve fibers themselves. Different brain regions can be determined from structures in the measured image stacks and in the derived parameter maps. A segmentation of brain regions is not aim of the reconstruction. We are only interested in the course, i.e. the orientation vectors of (crossing) nerve fibers in each image pixel (see also Response to #7).
