### 2020-10-14 3d-pli/SLIX#7 @miriammenzel
There exist different ways and existing tools how the parameter maps generated by SLIX can be further processed. The direction maps generated by SLIX contain the in-plane direction angles of the nerve fibers, which can be used as a-priori information for tractography algorithms or as direct input (in-plane vector maps) for further analysis. For example, the in-plane direction angles can be combined with out-of-plane inclination angles obtained from 3D-Polarized Light Imaging measurements of the same tissue sample to generate three-dimensional vectors in space. Using existing registration tools, the average intensities of consecutive  brain sections can be registered onto each other, hence creating a volume of (crossing) vectors. This set of unit vectors can be used as direct input for streamline-based tractography algorithms, as described in https://doi.org/10.1007/978-3-658-25326-4_17 . The input needed depends on the implementation of the tractography algorithm, and there exist already other registration and tractography tools. Therefore, SLIX provides the direction maps (and other parameter maps) to keep the software as focused as possible.
