### 2020-10-14 3d-pli/SLIX#8 @miriammenzel
We agree that we can explain the format of the input data in more detail:
The images from the microscope (raw format) are converted to .nii or .tiff images. The input is an SLI image stack that contains a series of these images (.nii or .tiff), where each image was recorded with a different illumination angle (from Phi=0° to 360°, starting on top and rotating clock-wise, as shown in Figure 2 in the arXiv paper). The software assumes that the measurement has been performed with equidistant illumination angles over a full range of 360°. The number of images defines the illumination angles (no metadata needed). In your example, the images represent measurements with 24 different illumination angles (Phi=0°, 15°, …, 345°). The first and second dimensions are the number of image pixels in x and y, the third dimension is the number of images, the forth dimension (in the .nii image) is ignored.
We will include a more detailed description of the data format in the README.
