### 2020-10-14 3d-pli/SLIX#6 @miriammenzel
We agree that we should motivate the use of 8% prominence. We will include a more detailed explanation (and a reference to Appx. A) in the Summary of the JOSS paper and in the “Introduction” of the README.
The derivation of the 8% value is a major part of the arXiv paper and cannot be explained here in full detail. To obtain this value, many different brain regions with (crossing) nerve fibers and well-known fiber orientations have been considered. In this way, we could find the best compromise between correctly and wrongly detected peaks. Usually, we do not know the true underlying fiber orientations of the tissue. Therefore, it is important to know for which prominence the peaks are indicators of the fiber orientations, and for which prominence the peaks are caused by noise or details in the fiber structure that are not of interest. This cannot be determined on a small section of the data.
The value of 8% prominence has been shown to be the optimal choice for the generation of reliable fiber orientations. If you like, we can add an optional parameter that allows experienced users to change this value? (although we would not recommend this).

You are right that the signal amplitude (I_max – I_min) may vary across setups. Therefore, the prominence is defined as a percentage of the signal amplitude (after normalizing the signal by the average value, i.e. accounting for different input intensities).

It is not possible to compute a goodness-of-fit because the SLI profiles are very different in shape and cannot be fitted by a reasonable function without overfitting. However, the parameter maps show indirectly how reliable the results are: The “low_prominence_peaks” map shows the number of peaks with a low prominence; the “high_prominence_peaks” map shows the number of peaks with a high prominence. In addition, the “peakprominence” map indicates the average prominence of the peaks. Regions with a clear signal, i.e. a well-defined fiber orientation, show only prominent peaks, a high average peak prominence, a small peak width, and a large peak distance. At the same time, the parameter maps can also be used to determine regions with out-of-plane fibers (regions with one prominent peak, large peak width, and lower peak prominence), as described in more detail in the arXiv paper.

Walking the user through the process from the raw data to the output is a good idea. We will add a more detailed description to the “Example” and “Resulting parameter maps” sections in the README, describing all parameter maps that can be generated with SLIX. We can include a Wiki and/or Jupyter notebook tutorial if you like?

The parameter “--with_mask” (and “--mask_threshold”) can be used to set all image pixels with low signals to zero, i.e. they are considered as background and not further evaluated. The parameter “--with_smoothing” can be used when the measurement has been performed with small steps of illumination angles (< 5°) to smoothen the resulting SLI signals and reduce the impact of irrelevant details in the fiber structure (cf. orange vs. black curve in Figure 1a in the JOSS paper). We will add more detailed descriptions to the “Optional parameters” section of the README.
