### 2020-10-14 3d-pli/SLIX#6 @alexrockhill
Thanks for your thoughtful response.

That makes sense about referencing the paper for the details of the 8% and other parameters but for just high-level user functionality, I would imagine that a user might at some point have an artifactually bright pixel that may throw off the stability of these metrics and they may need to change that value.

I think references related to when a user would actually use the data and what their data would look like when they should do so would be very helpful (i.e. the figure 1a reference).
