### 2020-10-14 3d-pli/SLIX#4 @alexrockhill
That all makes more sense, thanks for the explanation. I think the example could be greatly improved by walking the user through what you've just explained, whereas now they are just given the outputs without explanation.

Re the orientation vectors vs reconstructed fibers, I agree that this ties in with #7 and, I don't mean to be overly constrained toward the TRACULA approach to analysis as from my background that is what I know best, but I think it could be clarified by an example and/or more documentation what to do with the output to get to a place where a more diverse set of users could understand what to do and what software packages or what needs to be done that would help them continue their analysis.
