### 2020-10-14 3d-pli/SLIX#6 @Thyre
I agree that some users may want to change these high values if a pixel is artificially bright or even out of curiosity. Currently, users would only be able to change this value by modifying the Python file itself, which is especially impractical when SLIX is installed as a Python package.
I will add parameters to change these values at startup from the command line. This should be done by tomorrow.

I also agree that some examples of the parameters and when they are used could be added to our README file. We will add further explanations.
