### 2020-10-14 4f536f0
Update README (parameter maps, TOC)
