+++ README.md
@@ -1,8 +1,45 @@
-# SLIX - Scattered Light Imaging ToolboX
+# SLIX &ndash; Scattered Light Imaging ToolboX
 
 ![https://jugit.fz-juelich.de/j.reuter/slix/-/raw/assets/SLIX_Logo.png](https://jugit.fz-juelich.de/j.reuter/slix/-/raw/assets/SLIX_Logo.png)
 
-# Introduction 
+
+<!-- @import "[TOC]" {cmd="toc" depthFrom=2 depthTo=6 orderedList=false} -->
+
+<!-- code_chunk_output -->
+
+- [Introduction](#introduction)
+- [How to clone SLIX](#how-to-clone-slix)
+- [How to install SLIX as Python package](#how-to-install-slix-as-python-package)
+- [`SLIXParameterGenerator`](#slixparametergenerator)
+  - [Required parameters](#required-parameters)
+  - [Optional parameters](#optional-parameters)
+  - [Output](#output)
+  - [Example](#example)
+    - [How to run the demo yourself:](#how-to-run-the-demo-yourself)
+      - [1. Download the needed files:](#1-download-the-needed-files)
+      - [2. Run SLIX:](#2-run-slix)
+  - [Resulting parameter maps](#resulting-parameter-maps)
+      - [Average](#average)
+      - [Low Prominence Peaks](#low-prominence-peaks)
+      - [High Prominence Peaks](#high-prominence-peaks)
+      - [Average Peak Prominence](#average-peak-prominence)
+      - [Average Peak Width](#average-peak-width)
+      - [Peak Distance](#peak-distance)
+      - [Direction Angles](#direction-angles)
+        - [Maximum](#maximum)
+        - [Minimum](#minimum)
+- [Additional tools](#additional-tools)
+  - [`SLIXLineplotParameterGenerator`](#slixlineplotparametergenerator)
+- [Performance metrics](#performance-metrics)
+- [Authors](#authors)
+- [References](#references)
+- [Acknowledgements](#acknowledgements)
+- [License](#license)
+
+<!-- /code_chunk_output -->
+
+
+## Introduction 
 
 *Scattered Light Imaging (SLI)* is a novel neuroimaging technique that resolves the substructure of nerve fibers, especially in regions with crossing nerve fibers, in whole brain sections with micrometer resolution ([Menzel et al. (2020b)](https://arxiv.org/abs/2008.01037)). The measurement principle was first introduced by [Menzel et al. (2020a)](http://dx.doi.org/10.1103/PhysRevX.10.021002): A histological brain section is illuminated under oblique incidence of light from different angles. The measurement is performed with a constant polar angle of illumination and different azimuthal angles (*directions of illumination* <img src="https://render.githubusercontent.com/render/math?math=\phi">). For each direction of illumination, the intensity of light that is transmitted under normal incidence is recorded. The resulting images form a series (*SLI image stack*) in which each image pixel contains a light intensity profile (*SLI profile* <img src="https://render.githubusercontent.com/render/math?math=I(\phi)">).
 
@@ -98,17 +135,74 @@ The execution of both commands should take around one minute max. The resulting 
 
 ### Resulting parameter maps
 
[…]
