+++ README.md
@@ -3,34 +3,21 @@
 ![https://jugit.fz-juelich.de/j.reuter/slix/-/raw/assets/SLIX_Logo.png](https://jugit.fz-juelich.de/j.reuter/slix/-/raw/assets/SLIX_Logo.png)
 
 
-<!-- @import "[TOC]" {cmd="toc" depthFrom=2 depthTo=6 orderedList=false} -->
+<!-- @import "[TOC]" {cmd="toc" depthFrom=2 depthTo=3 orderedList=false} -->
 
 <!-- code_chunk_output -->
 
 - [Introduction](#introduction)
-- [How to clone SLIX](#how-to-clone-slix)
-- [How to install SLIX as Python package](#how-to-install-slix-as-python-package)
-- [`SLIXParameterGenerator`](#slixparametergenerator)
-  - [Required parameters](#required-parameters)
-  - [Optional parameters](#optional-parameters)
-  - [Output](#output)
+- [Installation](#installation)
+- [Generation of Parameter Maps](#generation-of-parameter-maps)
+  - [Required Arguments](#required-arguments)
+  - [Optional Arguments](#optional-arguments)
   - [Example](#example)
-    - [How to run the demo yourself:](#how-to-run-the-demo-yourself)
-      - [1. Download the needed files:](#1-download-the-needed-files)
-      - [2. Run SLIX:](#2-run-slix)
-  - [Resulting parameter maps](#resulting-parameter-maps)
-      - [Average](#average)
-      - [Low Prominence Peaks](#low-prominence-peaks)
-      - [High Prominence Peaks](#high-prominence-peaks)
-      - [Average Peak Prominence](#average-peak-prominence)
-      - [Average Peak Width](#average-peak-width)
-      - [Peak Distance](#peak-distance)
-      - [Direction Angles](#direction-angles)
-        - [Maximum](#maximum)
-        - [Minimum](#minimum)
-- [Additional tools](#additional-tools)
-  - [`SLIXLineplotParameterGenerator`](#slixlineplotparametergenerator)
-- [Performance metrics](#performance-metrics)
+  - [Resulting Parameter Maps](#resulting-parameter-maps)
+- [Evaluation of SLI Profiles](#evaluation-of-sli-profiles)
+  - [Required Arguments](#required-arguments-1)
+  - [Optional Arguments](#optional-arguments-1)
+- [Performance Metrics](#performance-metrics)
 - [Authors](#authors)
 - [References](#references)
 - [Acknowledgements](#acknowledgements)
@@ -45,7 +32,8 @@
 
 This repository contains the *Scattered Light Imaging ToolboX (SLIX)* - an open-source Python package that allows a fully automated evaluation of SLI measurements and the generation of different parameter maps. For a given SLI image stack, `SLIXParameterGenerator` is able to compute up to 12 (8 + 4 optional) parameter maps providing different information about the measured brain tissue sample, e.g. the individual in-plane direction angles of the nerve fibers for regions with up to three crossing nerve fiber bundles. Individual parameter maps can be selected through command line parameters. With `SLIXLineplotParameterGenerator`, it is possible to use existing SLI profiles (txt-files with a list of intensity values) as input and compute the corresponding parameter set (txt-file) for each SLI profile, which contains the number of peaks, the position (<img src="https://render.githubusercontent.com/render/math?math=\phi">) of the maximum and minimum, and the peak positions.
 
-## How to clone SLIX
+## Installation 
+##### How to clone SLIX
 ```
 git clone git@github.com:3d-pli/SLIX.git
 cd SLIX
@@ -57,7 +45,7 @@ source venv/bin/activate
 pip3 install -r requirements.txt
[…]
