### 2020-10-15 d7f6ce2
Update Paper (modified Figure 1, add more details)
