+++ paper/paper.md
@@ -45,6 +45,6 @@ Figure \ref{figure2}(a-h) shows parameter maps generated with `SLIX` for two art
 The authors thank Markus Axer and Katrin Amunts for their continuous support. 
 This work was funded by the Helmholtz Association port-folio theme "Supercomputing and Modeling for the Human Brain" and by the European Union's Horizon 2020 Framework Programme for Research and Innovation under the Specific Grant Agreement No. 785907 and 945539 ("Human Brain Project" SGA2 and SGA3).
 
-![Parameter maps generated with `SLIX`, shown exemplary for two artificially crossing sections of human optic tracts: (a) average of the line profiles; (b/c) number of all/prominent peaks in the line profiles; (d) average prominence of the peaks in the line profiles, normalized by the average of the line profile; (e) average width of all prominent peaks in the line profiles; (f) distance between two prominent peaks; (g/h) in-plane direction angles of the nerve fibers (two out of three possible directions); (i) visualization of direction angles in g and h, showing the crossing nerve fibers in the center. This figure has been adapted from @menzel2020-NeuroImage, Figure 8. Subfigures a-h were generated with the SLIX software, using a viridis color map to display the results and manually masking the tissue regions. Subfigure b shows the number of all peaks, i.e. the sum of low and high prominence peaks. Subfigure i was generated from the results in g,h by representing the direction angles for 24x24 pixels by a line with the respective polar/direction angle (not part of the software).  \label{figure2}](figure2.pdf){width=100%}
+![Parameter maps generated with `SLIX`, shown exemplary for two artificially crossing sections of human optic tracts: (a) average of the line profiles; (b/c) number of all/prominent peaks in the line profiles; (d) average prominence of the peaks in the line profiles, normalized by the average of the line profile; (e) average width of all prominent peaks in the line profiles; (f) distance between two prominent peaks; (g/h) in-plane direction angles of the nerve fibers (two out of three possible directions); (i) visualization of direction angles in g and h, showing the crossing nerve fibers in the center. This figure has been adapted from @menzel2020-NeuroImage, Figure 8. Subfigures a-h were generated with the SLIX software, using a viridis color map to display the results and manually masking the tissue regions. Subfigure b shows the number of all peaks, i.e. the sum of low and high prominence peaks. Subfigure i was generated from the results in g,h by representing the direction angles for 24x24 pixels by a line with the respective polar/direction angle (not part of the software).  \label{figure2}](figure2.pdf){width=80%}
 
 # References
