+++ paper/paper.md
@@ -24,27 +24,25 @@ bibliography: paper.bib
 # Summary
 
 To study the structure and function of the brain, a detailed reconstruction of the brain's nerve fiber architecture is required. Many neuroimaging techniques, however, struggle with the correct reconstruction of crossing nerve fibers. 
-The recently developed technique *Scattered Light Imaging (SLI)* resolves the substructure of (crossing) nerve fibers in whole brain sections with micrometer resolution [@menzel2020-NeuroImage]. The measurement principle was first introduced by @menzel2020. Unstained histological brain sections are illuminated from different angles and the transmitted (scattered) light behind the sample is measured under normal incidence (see Figure \ref{figure}(a)). Every pixel in the resulting SLI image series contains a light intensity profile (*SLI profile* I(<img src="https://render.githubusercontent.com/render/math?math=\phi">), cf. Figure \ref{figure}(b,c)) which is characteristic for the brain tissue structure at this point. The number and position of significant peaks in the line profiles can be used, for example, to determine the individual orientations of (crossing) nerve fibers in each image pixel. 
+The recently developed technique *Scattered Light Imaging (SLI)* allows to explore the substructure of (crossing) nerve fibers in whole brain sections with micrometer resolution, by providing crucial information such as the orientations of crossing nerve fibers in a measured image pixel [@menzel2020-NeuroImage]. The measurement principle was first introduced by @menzel2020. Unstained histological brain sections are illuminated from different angles and the transmitted (scattered) light behind the sample is measured under normal incidence (see Figure \ref{figure}a). Every pixel in the resulting SLI image series contains a light intensity profile (*SLI profile* $I(\phi)$), cf. Figure \ref{figure}b-c) which is characteristic for the brain tissue structure at this point. The number and position of significant peaks in the SLI profiles can be used, for example, to determine the individual orientations of (crossing) nerve fibers in each image pixel. 
 
-Here, we present the *Scattered Light Imaging ToolboX* (`SLIX`) - an open-source Python package that allows for the first time a fully automated evaluation of the SLI measurements and the generation of different parameter maps (see Figure \ref{figure2}) which reflect different characteristics of the SLI profiles: the average, the number of peaks, the average peak prominence, the average peak width, the distance between peaks, and the position of the peaks from which the in-plane direction angles of the nerve fibers are computed in regions with up to three crossing nerve fiber bundles. The user can choose which parameter maps are generated by using specific command line parameters.
+Here, we present the *Scattered Light Imaging ToolboX* (`SLIX`) - an open-source Python package that allows for the first time a fully automated evaluation of SLI measurements and the generation of different parameter maps (see Figure \ref{figure2}) which reflect different characteristics of the SLI profiles: the average, the number of peaks, the average peak prominence, the average peak width, the distance between peaks, and the position of the peaks from which the in-plane direction angles of the nerve fibers are computed in regions with up to three crossing nerve fiber bundles. The user can choose which parameter maps are generated by using specific command line parameters.
 
[…]
