### 2020-10-15 3d-pli/SLIX#4 @matteomancini
I ran the first example as well, but without downsampling, trying to replicate the figures in the section "Resulting parameter maps". However, for some reason the results are different from the ones showed in the README (screenshot for direction 1 attached).
I tried both on macOS Catalina and on Ubuntu 18.04, and the results are the same. If it a visualisation issue, is it possible to advise for a suitable image viewer?
Here are the actual maps generated: [Archive.zip](https://github.com/3d-pli/SLIX/files/5383990/Archive.zip)
On a related note, I wanted to say that I really appreciate the recent inclusion of the progress bar! 👍
