### 2020-10-15 3d-pli/SLIX#10 @matteomancini
[DOC] SLIXLineplotParameterGenerator example
#Is it possible to add an example for SLIXLineplotParameterGenerator (even a very brief one) based on one of the two datasets used for SLIXParameterGenerator and providing a text file with intensity values? This would make easier to try out that specific functionality.

https://github.com/openjournals/joss-reviews/issues/2675
