+++ paper/paper.md
@@ -37,12 +37,10 @@ For a full documentation of `SLIX`, the reader is referred to our [GitHub page](
 
 ![(a) Schematic drawing of the SLI measurement: The sample is illuminated from different angles (with constant polar angle $\theta$ and different equidistant azimuthal angles $\phi$) and the transmitted light behind the sample is measured under normal incidence. Each pixel in the resulting image series contains a characteristic light intensity profile (SLI profile $I(\phi)$). (b) SLI profile measured in steps of 15°. The prominence of the peaks (in red) is computed by the difference between the top of the peak and the highest of the two neighboring minima. The peak width (dark blue) is determined as the full width of the peak at a height corresponding to the peak height minus half of the peak prominence. The determined positions of the peaks (vertical lines) have been slightly corrected by computing the centroid of the peak tip as described in @menzel2020-NeuroImage, Appx. B to account for discretization artifacts. Only *prominent* peaks, i.e. peaks with a prominence above 8% of the total signal amplitude (max - min), are used for further evaluation (green/magenta lines). Peaks with lower prominences are expected to be caused by noise or details in the fiber structure that are not of interest (for derivation see @menzel2020-NeuroImage, Appx. A). The fiber direction angles $\varphi_1$ and $\varphi_2$ are computed from the mid positions of the prominent peak pairs. (c)  SLI profile measured in steps of 1° (orange: original line profile, black: smoothed line profile).\label{figure}](figure_.pdf){width=100%}
 
-Figure \ref{figure2}a-h shows parameter maps generated with `SLIX` for two artificially crossing sections of human optic tracts. The sections are delineated by black contour lines for better reference. 
-
 # Acknowledgements
 The authors thank Markus Axer and Katrin Amunts for their continuous support. 
 This work was funded by the Helmholtz Association port-folio theme "Supercomputing and Modeling for the Human Brain" and by the European Union's Horizon 2020 Framework Programme for Research and Innovation under the Specific Grant Agreement No. 785907 and 945539 ("Human Brain Project" SGA2 and SGA3).
 
-![Parameter maps generated with `SLIX`, shown exemplary for two artificially crossing sections of human optic tracts: (a) average of the line profiles; (b/c) number of all/prominent peaks in the line profiles; (d) average prominence of the peaks in the line profiles, normalized by the average of the line profile; (e) average width of all prominent peaks in the line profiles; (f) distance between two prominent peaks; (g/h) in-plane direction angles of the nerve fibers (two out of three possible directions); (i) visualization of direction angles in g and h, showing the crossing nerve fibers in the center. This figure has been adapted from @menzel2020-NeuroImage, Figure 8. Subfigures a-h were generated with `SLIX`, using a viridis color map to display the results and manually masking the tissue regions. Subfigure b shows the number of all peaks, i.e. the sum of low and high prominence peaks. Subfigure i was generated from the results in g,h by representing the direction angles of 24x24 pixels by a line with the respective polar/direction angle (not part of the software).  \label{figure2}](figure2.pdf){width=90%}
[…]
