### 2020-10-15 3d-pli/SLIX#4 @miriammenzel
@matteomancini: The generated parameter maps look good (how long did it take for you to run the example with full resolution?)
Which image viewer did you use? It seems that the image does not show the full range of values (-1 is displayed in dark gray values, all other values are shown in white). Examples for suitable image viewers are ImageJ or FIJI (https://imagej.net/Fiji). We will add a note in the README.
