+++ README.md
@@ -9,14 +9,15 @@
 
 - [Introduction](#introduction)
 - [Installation](#installation)
-- [Generation of Parameter Maps](#generation-of-parameter-maps)
+- [Evaluation of SLI Profiles](#evaluation-of-sli-profiles)
   - [Required Arguments](#required-arguments)
   - [Optional Arguments](#optional-arguments)
   - [Example](#example)
-  - [Resulting Parameter Maps](#resulting-parameter-maps)
-- [Evaluation of SLI Profiles](#evaluation-of-sli-profiles)
+- [Generation of Parameter Maps](#generation-of-parameter-maps)
   - [Required Arguments](#required-arguments-1)
   - [Optional Arguments](#optional-arguments-1)
+  - [Example](#example-1)
+  - [Resulting Parameter Maps](#resulting-parameter-maps)
 - [Performance Metrics](#performance-metrics)
 - [Authors](#authors)
 - [References](#references)
@@ -53,9 +54,51 @@ cd SLIX
 python3 setup.py install
 ```
 
+## Evaluation of SLI Profiles
+
+`SLIXLineplotParameterGenerator` allows the evaluation of individual SLI profiles (txt-files with a list of intensity values): For each SLI profile, the maximum, minimum, number of prominent peaks, corrected peak positions, fiber direction angles, and average peak prominence are computed and stored in a text file. The user has the option to generate a plot of the SLI profile that shows the profile together with the peak positions before/after correction. The corrected peak positions are determined by calculating the geometric center of the peak tip, improving the accuracy of the determined peak positions in discretized SLI profiles. If not defined otherwise, the peak tip height corresponds to 6% of the total signal amplitude (for derivation, see [Menzel et al. (2020b)](https://arxiv.org/abs/2008.01037), Appx. B).
+
+```
+SLIXLineplotParameterGenerator -i [INPUT-TXT-FILES] -o [OUTPUT-FOLDER] [[parameters]]
+```
+### Required Arguments
+| Argument      | Function                                                                    |
+| ------------------- | --------------------------------------------------------------------------- |
+| `-i, --input`  | Input text files, describing the SLI profiles (list of intensity values). |
+| `-o, --output` | Output folder used to store the characteristics of the SLI profiles in a txt-file (Max, Min, Num_Peaks, Peak_Pos, Directions, Prominence). Will be created if not existing. |
+
+### Optional Arguments
+| Argument      | Function                                                                    |
+| -------------- | --------------------------------------------------------------------------- |
+| `--smoothing`  | Smoothing of SLI profiles before evaluation. The smoothing is performed using a Savitzky-Golay filter with 45 sampling points and a second order polynomial. (Designed for measurements with <img src="https://render.githubusercontent.com/render/math?math=\Delta\phi"> < 5° steps.) |
+| `--with_plots` | Generates plots (png-files) showing the SLI profiles and the determined peak positions (orange dots: before correction; green crosses: after correction). |
+| `--target_peak_height` | Change peak tip height used for correcting the peak positions. (Default: 6% of total signal amplitude). Only recommended for experienced users! |
+
+### Example
+The following example demonstrates the evaluation of two SLI profiles, which can be found in the "examples" folder of the SLIX repository:
+```
+SLIXLinePlotParameterGenerator -i examples/*.txt -o output --with_plots
+```
+The resulting plot and txt-file are shown below, exemplary for one of the SLI profiles (90-Stack-1647-1234.txt):
+
+<img src="https://jugit.fz-juelich.de/j.reuter/slix/-/raw/assets/90-Stack-1647-1234.png" height="327">
+
[…]
