### 2020-10-15 3d-pli/SLIX#4 @matteomancini
Thanks @miriammenzel it makes sense, I was using normal image viewers (e.g. Preview on macOS). I did not recorded execution time, but it was more or less 1 hour on macOS (2.8 GHz Quad-Core Intel Core i7) and around 30 minutes on Ubuntu (2.1GHz 8 Core Intel Xeon Silver 4110).
