### 2020-10-15 3d-pli/SLIX#7 @matteomancini
I agree with @alexrockhill - as someone who works with tractography but is not that familiar with light imaging it is not immediately clear to me how to practically use the output of SLIX to inform a tracking algorithm implemented in some software (e.g. MRtrix, Diffusion Toolkit). I think that something that could help to immediately make the connection would be some functionality to either generate automatically the vectors represented in figure 2i (as it was mentioned in #4 - ideally in Python) or to show how to use one of the maps available at the moment as an input to an existing tracking tool.

@miriammenzel I think that the procedure you describe is very interesting, but also has some challenges, as registering contiguous sections to reconstruct a 3D volume leads to several issues (e.g. "banana" effect, z-shift). Of course these go beyond the scope of SLIX and the JOSS paper, but this is to say the overall procedure could quickly become quite complex. Again, I do not believe such an elaborate example (involving registration and tracking) is required to present an interesting tool such as SLIX. What would make it feel more complete is explaining how one of the outputs could be easily used to inform tracking procedures.
