### 2020-10-15 #2675 @matteomancini
I had some time to go through the SLIX repo and I think it is overall a very nice tool. It was very easy to have it up and running, and the command line interface is quite immediate. I had some visualization issue but it was because of what I was using to view the maps and the authors were very quick to reply. One suggestion that I have is to upload the package to PyPI: the installation is already quite simple, but it would be even easier to just type `pip install slix`.

I think the main thing missing (to fully check both the substantial scholarly effort and the functionality documentation) is an explicit explanation of how to use the output maps to inform tracking procedures in order to e.g. reconstruct streamlines. This was brought up by @alexrockhill here: https://github.com/3d-pli/SLIX/issues/7 – I think that a full tutorial on how to do some sort of “SLI-based tractography” is outside the scope of this work, but it would be appropriate to practically explain how to use one of the maps in a tracking scenario. Also generating automatically the vectorial data showed in figure 2i with Python would be a good way to bridge to tracking.

Another very minor issue I mentioned in https://github.com/3d-pli/SLIX/issues/10 is to add a practical example of how to use the other main CLI tool, SLIXLineplotParameterGenerator. Apart from that, I think the README is quite detailed and the methods are well commented.

Finally, the paper is well written, and I think that the current functionality summary and the statement of need are concise and appropriate. The only thing missing (but it may be due to my lack of experience with light imaging literature) is any mention of other similar tools with the related references. It may be the case that SLIX is the first of its kind for Scattered Light Imaging, but if it is the case a couple of sentences could be spent on similar tools for collateral techniques (e.g. polarized light imaging).

In any case, I want to stress that it is a very interesting tool!
