### 2020-10-15 3d-pli/SLIX#6 @Thyre
I‘ve added the parameter „—prominence_threshold“ which allows the user to change the prominence used for the peak detection (https://github.com/3d-pli/SLIX/tree/6-method-explanations). The update to the README.md is still in progress (https://github.com/3d-pli/SLIX/tree/update-readme) but will include a more detailed explanation about the parameters and their function.
