### 2020-10-15 3d-pli/SLIX#10 @Thyre
I absolutely agree that we should include one or two examples for `SLIXLineplotParameterGenerator` in the README.md. I've added two example files in [10-SLIXLineplotParameterGenerator-example](https://github.com/3d-pli/SLIX/tree/10-SLIXLineplotParameterGenerator-example). The README.md is still in progress but will contain both instructions on how to use `SLIXLineplotParameterGenerator` and what results are being generated when running the program for the given examples (see [update-readme](https://github.com/3d-pli/SLIX/tree/update-readme)).
