### 2020-10-15 3d-pli/SLIX#11 @Thyre
Publish package on PyPi
As @matteomancini mentions in the [review for the JOSS paper](https://github.com/openjournals/joss-reviews/issues/2675), the installation process could be simplified for many Python users by uploading the SLIX package to PyPi. There are several advantages such as a one-line installation without manual download or a simpler update process.

I haven't actually released anything in PyPi yet and I don't know how to do it, but I will definitely look at it, as it would be a great improvement for the users and it might also make the package easier to find.
