### 2020-10-15 1a9d397
Update README (more detailed Introduction, with figure)
