+++ paper/paper.bib
@@ -21,3 +21,14 @@
   Owner                    = {mmenzel},
   Timestamp                = {2018.04.16}
 }
+
+@Article{MAxer2011_1,
+  Title                    = {A novel approach to the human connectome: Ultra-high resolution mapping of fiber tracts in the brain},
+  Author                   = {Axer, M. and Amunts, K. and Gr{\"a}ssel, D. and Palm, C. and Dammers, J. and Axer, H. and Pietrzyk, U. and Zilles, K.},
+  Journal                  = {Neuro{I}mage},
+  Year                     = {2011},
+  Number                   = {2},
+  Pages                    = {1091--1101},
+  Volume                   = {54}
+}
+
