### 2020-10-15 50ce844
Update paper (reference, bib)
