+++ paper/paper.bib
@@ -32,3 +32,13 @@
   Volume                   = {54}
 }
 
+@Article{MAxer2011_2,
+  Title                    = {High-resolution fiber tract reconstruction in the human brain by means of three-dimensional polarized light imaging},
+  Author                   = {Axer, M. and Gr{\"a}ssel, D. and Kleiner, M. and Dammers, J. and Dickscheid, T. and Reckfort, J. and H{\"u}tz, T. and Eiben, B. and Pietrzyk, U. and Zilles, K. and Amunts, K.},
+  Journal                  = {Front. Neuroinform.},
+  Year                     = {2011},
+  Number                   = {34},
+  Pages                    = {1--13},
+  Volume                   = {5}
+}
+
