+++ paper/paper.md
@@ -23,7 +23,7 @@ bibliography: paper.bib
 
 # Summary
 
-To study the structure and function of the brain, a detailed reconstruction of the brain's nerve fiber architecture is required. However, many neuroimaging techniques, such as 3D-Polarized Light Imaging [@MAxer2011_1], struggle with the correct reconstruction of crossing nerve fibers. 
+To study the structure and function of the brain, a detailed reconstruction of the brain's nerve fiber architecture is required. However, many neuroimaging techniques, such as 3D-Polarized Light Imaging [@MAxer2011_1,@MAxer2011_2], struggle with the correct reconstruction of crossing nerve fibers. 
 The recently developed technique *Scattered Light Imaging (SLI)* allows to explore the substructure of (crossing) nerve fibers in whole brain sections with micrometer resolution, by providing crucial information such as the orientations of crossing nerve fibers in a measured image pixel [@menzel2020-NeuroImage]. The measurement principle was first introduced by @menzel2020. Unstained histological brain sections are illuminated from different angles and the transmitted (scattered) light behind the sample is measured under normal incidence (see Figure \ref{figure}a). Every pixel in the resulting SLI image series contains a light intensity profile (*SLI profile* $I(\phi)$, cf. Figure \ref{figure}b-c) which is characteristic for the brain tissue structure at this point. The number and position of significant peaks in the SLI profiles can be used, for example, to determine the individual orientations of (crossing) nerve fibers in each image pixel. 
 
 Here, we present the *Scattered Light Imaging ToolboX* (`SLIX`) &ndash; an open-source Python package that allows for the first time a fully automated evaluation of SLI measurements and the generation of different parameter maps (see Figure \ref{figure2}a-h) which reflect different characteristics of the SLI profiles: the average, the number of peaks, the average peak prominence, the average peak width, the distance between peaks, and the position of the peaks from which the in-plane direction angles of the nerve fibers are computed in regions with up to three crossing nerve fiber bundles. The user can choose which parameter maps are generated by using specific command line parameters. The resulting direction maps can be used to visualize the individual orientations of in-plane (crossing) nerve fibers, as shown in Figure \ref{figure2}i.
@@ -41,6 +41,8 @@ For a full documentation of `SLIX`, the reader is referred to our [GitHub page](
 The authors thank Markus Axer and Katrin Amunts for their continuous support. 
 This work was funded by the Helmholtz Association port-folio theme "Supercomputing and Modeling for the Human Brain" and by the European Union's Horizon 2020 Framework Programme for Research and Innovation under the Specific Grant Agreement No. 785907 and 945539 ("Human Brain Project" SGA2 and SGA3).
 
+# References
+
 ![Parameter maps generated with `SLIX`, shown exemplary for two artificially crossing sections of human optic tracts: (a) average intensity in the SLI profiles; (b/c) number of all/prominent peaks in the SLI profiles; (d) average prominence of the peaks in the SLI profiles, normalized by the average of the profile; (e) average width of all prominent peaks in the SLI profiles; (f) distance between two prominent peaks; (g/h) in-plane direction angles of the nerve fibers (two out of three possible directions); (i) visualization of direction angles in g and h, showing the crossing nerve fibers in the center. This figure has been adapted from @menzel2020-NeuroImage, Figure 8. Subfigures a-h were generated with `SLIX`, using a viridis color map to display the results and manually masking the tissue regions. Subfigure b shows the number of all peaks, i.e. the sum of low and high prominence peaks. Subfigure i was generated from the results in g,h by representing the direction angles of 24x24 pixels by a line with the respective polar/direction angle (not part of the software).  \label{figure2}](figure2.pdf){width=90%}
 
-# References
+
