### 2020-10-15 660d77c
Update of README and Paper, Introduction of new parameters for both command line tools. (#12)
* Fixed some minor bugs in SLIXLineplotParameterGenerator.

Sometimes the peak position was not correct resulting in wrong directions in the output file. Added peakprominence and direction to the output file

* Added two example files for SLIXLineplotParameterGenerator

* Update README (Evaluation of SLI Profiles, Example)

* Added prominence and peak height threshold

* Update paper (3D-PLI, bib)

Co-authored-by: mmenzel <m.menzel@fz-juelich.de>
