+++ paper/paper.bib
@@ -21,3 +21,24 @@
   Owner                    = {mmenzel},
   Timestamp                = {2018.04.16}
 }
+
+@Article{MAxer2011_1,
+  Title                    = {A novel approach to the human connectome: Ultra-high resolution mapping of fiber tracts in the brain},
+  Author                   = {Axer, M. and Amunts, K. and Gr{\"a}ssel, D. and Palm, C. and Dammers, J. and Axer, H. and Pietrzyk, U. and Zilles, K.},
+  Journal                  = {Neuro{I}mage},
+  Year                     = {2011},
+  Number                   = {2},
+  Pages                    = {1091--1101},
+  Volume                   = {54}
+}
+
+@Article{MAxer2011_2,
+  Title                    = {High-resolution fiber tract reconstruction in the human brain by means of three-dimensional polarized light imaging},
+  Author                   = {Axer, M. and Gr{\"a}ssel, D. and Kleiner, M. and Dammers, J. and Dickscheid, T. and Reckfort, J. and H{\"u}tz, T. and Eiben, B. and Pietrzyk, U. and Zilles, K. and Amunts, K.},
+  Journal                  = {Front. Neuroinform.},
+  Year                     = {2011},
+  Number                   = {34},
+  Pages                    = {1--13},
+  Volume                   = {5}
+}
+
