+++ examples/90-Stack-1647-1234.txt
@@ -1,0 +1,24 @@
+82
+90
+100
+99
+95
+93
+100
+115
+119
+105
+83
+78
+68
+74
+94
+90
+77
+75
+77
+79
+93
+86
+85
+73
