+++ README.md
@@ -1,14 +1,71 @@
-# SLIX - Scattered Light Imaging ToolboX
+# SLIX &ndash; Scattered Light Imaging ToolboX
 
 ![https://jugit.fz-juelich.de/j.reuter/slix/-/raw/assets/SLIX_Logo.png](https://jugit.fz-juelich.de/j.reuter/slix/-/raw/assets/SLIX_Logo.png)
 
-# Introduction 
 
-*Scattered Light Imaging (SLI)* is a novel neuroimaging technique that resolves the substructure of nerve fibers, especially in regions with crossing nerve fibers, in whole brain sections with micrometer resolution ([Menzel et al. (2020b)](https://arxiv.org/abs/2008.01037)). The measurement principle was first introduced by [Menzel et al. (2020a)](http://dx.doi.org/10.1103/PhysRevX.10.021002): A histological brain section is illuminated under oblique incidence of light from different angles. The measurement is performed with a constant polar angle of illumination and different azimuthal angles (*directions of illumination* <img src="https://render.githubusercontent.com/render/math?math=\phi">). For each direction of illumination, the intensity of light that is transmitted under normal incidence is recorded. The resulting images form a series (*SLI image stack*) in which each image pixel contains a light intensity profile (*SLI profile* <img src="https://render.githubusercontent.com/render/math?math=I(\phi)">).
+<!-- @import "[TOC]" {cmd="toc" depthFrom=2 depthTo=3 orderedList=false} -->
 
-This repository contains the *Scattered Light Imaging ToolboX (SLIX)* - an open-source Python package that allows a fully automated evaluation of SLI measurements and the generation of different parameter maps. For a given SLI image stack, `SLIXParameterGenerator` is able to compute up to 12 (8 + 4 optional) parameter maps providing different information about the measured brain tissue sample, e.g. the individual in-plane direction angles of the nerve fibers for regions with up to three crossing nerve fiber bundles. Individual parameter maps can be selected through command line parameters. With `SLIXLineplotParameterGenerator`, it is possible to use existing SLI profiles (txt-files with a list of intensity values) as input and compute the corresponding parameter set (txt-file) for each SLI profile, which contains the number of peaks, the position (<img src="https://render.githubusercontent.com/render/math?math=\phi">) of the maximum and minimum, and the peak positions.
+<!-- code_chunk_output -->
 
-## How to clone SLIX
+- [Introduction](#introduction)
+  - [SLI Measurement](#sli-measurement)
+  - [SLI Profiles](#sli-profiles)
+  - [Parameter Maps](#parameter-maps)
+- [Installation of SLIX](#installation-of-slix)
+- [Evaluation of SLI Profiles](#evaluation-of-sli-profiles)
+  - [Required Arguments](#required-arguments)
+  - [Optional Arguments](#optional-arguments)
+  - [Example](#example)
+- [Generation of Parameter Maps](#generation-of-parameter-maps)
+  - [Required Arguments](#required-arguments-1)
+  - [Optional Arguments](#optional-arguments-1)
+  - [Example](#example-1)
+  - [Resulting Parameter Maps](#resulting-parameter-maps)
+- [Performance Metrics](#performance-metrics)
+- [Authors](#authors)
+- [References](#references)
+- [Acknowledgements](#acknowledgements)
+- [License](#license)
+
+<!-- /code_chunk_output -->
+
+
+## Introduction 
+
[…]
