### 2020-10-15 3d-pli/SLIX#10 @miriammenzel
I have updated the [README.md](https://github.com/3d-pli/SLIX/blob/master/README.md) with a more detailed explanation and an example of how to use `SLIXLineplotParameterGenerator` in the Section "Evaluation of Line Profiles". Please let us know if we should add anything else there.
