### 2020-10-15 3d-pli/SLIX#8 @miriammenzel
I have added a more detailed description of the SLI measurement and the input data (SLI image stack) to the README.md [here](https://github.com/3d-pli/SLIX/blob/master/README.md#sli-measurement). I hope that - together with the illustration in the "Introduction" part - the sample data has become clear now.
