### 2020-10-15 3d-pli/SLIX#7 @miriammenzel
Thank you for your replies.
We are currently working on an automatic generation and visualization of the vector maps (with Python) – we will let you know as soon as this is done. The vector maps could be used as input e.g. for streamline tractography algorithms.
Unfortunately, we have not yet enough data to perform tractography on 3D volumes (we would need a sufficient number of consecutive brain sections measured with SLI that we can register onto each other to generate a 3D volume of vectors). This is part of future work (and will not change the scope of SLIX - as the tool does not include registration or tractography).
