### 2020-10-15 3d-pli/SLIX#6 @miriammenzel
The updated README.md now includes a more precise [description](https://github.com/3d-pli/SLIX/blob/master/README.md#optional-arguments-1) of the arguments `--with_mask` and `--with_smoothing` (with reference to figure 1a).
The [Introduction](https://github.com/3d-pli/SLIX/blob/master/README.md#introduction) has been completely revised, containing now details and explanations about the [SLI profiles](https://github.com/3d-pli/SLIX/blob/master/README.md#sli-profiles) and the relevant parameter choices (such as the 8% peak prominence or the 6% peak tip height).
As mentioned by @Thyre, the 8% prominence value can now be changed by the user (just as the 6% peak tip value).
In addition, I've added a thorough explanation of all generated [parameter maps](https://github.com/3d-pli/SLIX/blob/master/README.md#parameter-maps) (also below the "Example" section) and inserted figures for illustration purposes.
@alexrockhill: Please let us know if there is anything missing.
