### 2020-10-15 3d-pli/SLIX#4 @miriammenzel
The updated README.md now includes a thorough description and explanation of all resulting parameter maps and outputs ([here](https://github.com/3d-pli/SLIX/blob/master/README.md#resulting-parameter-maps)). We also added a revised version of the [paper](https://github.com/3d-pli/SLIX/blob/master/paper/paper.pdf), where we explain how the subfigures in Figure 2 were generated.

We are currently working on a visualization of the vector maps (with Python) – we will let you know as soon as this is done. We are also working on providing a brief tutorial to walk the user through all steps and provide different possibilities for visualization. We hope that this will clarify what the user can do with the output.
@alexrockhill @matteomancini: Would this meet your requirements?
