### 2020-10-15 #2675 @miriammenzel
> Finally, the paper is well written, and I think that the current functionality summary and the statement of need are concise and appropriate. The only thing missing (but it may be due to my lack of experience with light imaging literature) is any mention of other similar tools with the related references. It may be the case that SLIX is the first of its kind for Scattered Light Imaging, but if it is the case a couple of sentences could be spent on similar tools for collateral techniques (e.g. polarized light imaging).

@matteomancini: Thank you very much for your positive comment and your constructive feedback. Scattered Light Imaging is indeed a very novel technique. It has been recently developed by our group and first introduced in [Menzel et al. PRX 10:021002 (2020)](https://journals.aps.org/prx/abstract/10.1103/PhysRevX.10.021002). It is the first technique of its kind, using light scattering in the optical range to obtain information about neural organization in the brain. Therefore, there exist no similar tools to SLIX that we could mention here. I have added two references about polarized light imaging at the beginning of the [paper](https://github.com/3d-pli/SLIX/blob/master/paper/paper.pdf). Unfortunately, we already reached the maximum amount of words for the paper. Please let us know if we should still add more detail there (and what other information we could leave out instead).
Concerning the other things you mentioned (issues #7,10,11), we have already replied to them or are in the process of implementing them.
