### 2020-10-16 980977a
Update README (minor corrections)
