### 2020-10-16 4509bea
Fixed a bug in the toolbox where the program could crash if base parameters were used for some of the methods. None was replaced by numpy.inf to prevent this issue.
Added a first version of the visualization toolbox. Colors might be wrong and the vector directions have still to be verified by our internal tools but this first visualization is at least functional.

The current version will create the full image in Matplotlib and show it to the user. This might change in future versions.

Thinout is somewhat implemented but will be buggy. Please do not use that option.

Signed-off-by: Jan Andre 
[…]
