### 2020-10-16 c92d17e
Changed visualization to only write on the current axis.
This allows the user to add something to the image and showing the image
when he wants to.

Signed-off-by: Jan Andre Reuter <j.reuter@fz-juelich.de>
