### 2020-10-16 8631520
Fixed some bugs in toolbox.py. Added progress bars to some methods.
Signed-off-by: Jan Andre Reuter <jan.andre.reuter@hotmail.de>
