### 2020-10-16 d636110
Fixed some bugs in toolbox.py. Added progress bars to some methods.
Signed-off-by: Jan Andre Reuter <jan.andre.reuter@hotmail.de>
