### 2020-10-16 1469a5a
Added example for visualization with SLIX
Signed-off-by: Jan Andre Reuter <jan.andre.reuter@hotmail.de>
