+++ setup.py
@@ -6,7 +6,7 @@ setup(
     packages=['SLIX'],
     url='www.github.com/3d-pli/slix',
     license='MIT',
-    author='Jan André Reuter, Miriam Menzel',
+    author='Jan Andre Reuter, Miriam Menzel',
     author_email='j.reuter@fz-juelich.de',
     description='SLIX allows an automated evaluation of SLI measurements and generates different parameter maps.',
     scripts=['bin/SLIXParameterGenerator', 'bin/SLIXLineplotParameterGenerator'],
