### 2020-10-19 989ab1a
Added visualization of parameter maps.
Increased version number
Expanded example for visualization accordingly

Signed-off-by: Jan Andre Reuter <jan.andre.reuter@hotmail.de>
