### 2020-10-19 4a788b6
Expanded downsampling to work for multiple dimensions.
Replaced normal thinout in visualization of unit vectors by first using downsampling and using thinout after it to ensure that at vectors are present if at least 10% of the downsampled pixels are not BACKGROUND_COLOR.

Signed-off-by: Jan Andre Reuter <jan.andre.reuter@hotmail.de>
