+++ SLIX/visualization.py
@@ -1,25 +1,11 @@
 import numpy
 import pymp
 from matplotlib import pyplot as plt
+from PIL import Image
 import copy
+from . import toolbox
 
-CPU_COUNT = 8
-
-
-def downsample(image, sample_size=10):
-    # downsample image
-    x, y = image.shape
-    nx = numpy.ceil(x / sample_size).astype('int')
-    ny = numpy.ceil(y / sample_size).astype('int')
-    small_img = pymp.shared.array((nx, ny), dtype='float32')
-
-    with pymp.Parallel(CPU_COUNT) as p:
-        for i in p.range(0, nx):
-            for j in range(0, ny):
-                roi = image[sample_size * i:sample_size * i + sample_size, sample_size * j:sample_size * j + sample_size]
-                small_img[i, j] = numpy.round(numpy.median(roi))
-
-    return small_img
+CPU_COUNT = toolbox.CPU_COUNT
 
 
 def unit_vectors(directions):
@@ -37,7 +23,7 @@ def unit_vectors(directions):
         x-, y- and z-vector component in arrays
     """
     directions_rad = numpy.deg2rad(directions)
-    UnitX = numpy.sin(0.5 * numpy.pi) * numpy.cos(directions_rad)
+    UnitX = -numpy.sin(0.5 * numpy.pi) * numpy.cos(directions_rad)
     UnitY = numpy.sin(0.5 * numpy.pi) * numpy.sin(directions_rad)
 
     UnitX[numpy.isclose(directions, -1)] = 0
@@ -46,8 +32,37 @@ def unit_vectors(directions):
     return UnitX, UnitY
 
 
+def downsample(image, sample_size=10, background_value=-1):
+    # downsample image
+    if len(image.shape) == 2:
+        x, y = image.shape
+        z = 1
+    else:
+        x, y, z = image.shape
+
+    nx = numpy.ceil(x / sample_size).astype('int')
+    ny = numpy.ceil(y / sample_size).astype('int')
+    small_img = pymp.shared.array((nx, ny, z), dtype='float32')
+
+    for sub_image in range(z):
+        with pymp.Parallel(CPU_COUNT) as p:
+            for i in p.range(0, nx):
+                for j in range(0, ny):
+                    roi = image[sample_size * i:sample_size * i + sample_size,
+                                sample_size * j:sample_size * j + sample_size, sub_image]
+                    if numpy.count_nonzero(roi != background_value) > 0.1 * roi.size:
+                        small_img[i, j, sub_image] = numpy.median(roi[roi != background_value])
+                    else:
+                        small_img[i, j, sub_image] = background_value
+
+    if z == 1:
+        small_img = small_img.reshape((nx, ny))
+
+    return small_img
+
+
 def visualize_parameter_map(parameter_map, fig=None, ax=None, alpha=1, cmap='viridis', vmin=0, vmax=None):
-    if fig is None and ax is None:
+    if fig is None or ax is None:
         fig, ax = plt.subplots(1, 1)
 
     cmap_mod = copy.copy(plt.get_cmap(cmap))
@@ -62,24 +77,42 @@ def visualize_parameter_map(parameter_map, fig=None, ax=None, alpha=1, cmap='vir
 def visualize_unit_vectors(UnitX, UnitY, thinout=1, ax=None, alpha=0.8):
     if ax is None:
         ax = plt.gca()
-    skip = (slice(None, None, thinout), slice(None, None, thinout))
-    for i in range(UnitX.shape[-1]):
[…]
