### 2020-10-19 a0c79bd
SLIX/toolbox.py: Fixed a bug where the direction maps could not be reshaped with the toolbox.py as it had a second dimension with a length greather than 1
Fixed the visualization.py to correctly show "HSV" color maps without any problems. Fixed high and low ranged for normal parameter maps.

Added first version of Jupyter Notebook example. The example text is only in German for now but will be translated as soon as possible

Signed-off-by: Jan Andre Reuter <jan.andre.reuter@hotmail.de>
