### 2020-10-19 #2675 @alexrockhill
I largely agree with @matteomancini that SLIX serves a needed role in image processing for a new modality. My biggest critique that I have raised in issues thus far is that, although this software tool almost (directional plot PR pending AFAIK) replicates the authors' work on Bioarxiv, it doesn't seem as geared toward a new user who might have a slightly different setup and different needs as it could be. The authors have addressed these issues in the documentation, and I apologize for not responding sooner about their sufficiency, but I was planning to wait for @rly to finish the last review and then evaluate the process from start to finish again. I would finally add that the issue I raised about an example integration to a tractography software such as TRACULA and greater user control over parameters, are related to that, as a bit of a niche software (as opposed to a one-stop software), it could be made clearer where SLIX fits in a workflow that delivers a finished analysis. The main issue there being that there is more than one way to approach this image processing issue, and in my opinion, "Whether the software is sufficiently useful that it is likely to be cited by other researchers working in this domain" will depend on whether a researcher with good programming ability who collected this type of data would rather contribute to SLIX than start something new. I think SLIX is a package I personally would likely contribute to rather than starting something new if it were me, but more thorough documentation of why each analysis decision was taken in the software itself in a form that a user or developer could test and understand by working through themselves step-by-step would make this a non-decision. I don't think it's reasonable to expect SLIX to be a package with the stability and documentation of something like FSL or Freesurfer that have been around for decades in order to be published so I wouldn't set the bar at the issues brought up being completely resolved, but I think the groundwork and vision that the user base will expand and that SLIX could expand with it to meet an increasing diversity of needs would make SLIX do a better job of meeting this criteria. If it helps to have an example, I think this software tool https://pactools.github.io/ has a task that is similarly complex (there are multiple published ways to accomplish it), and it does a good job of walking a user/onboarding developer through step-by-step from data to results, has a flexible API and has the appearance, to me at least, that the methods could change or new methods could be integrated easily. Overall, I think the core algorithms seem well done and to do what is advertised and I think there is a SLIX-shaped hole in available software, but I wanted to ask the authors to consider a few comments about a more of a global perspective on SLIX's role in tractography software as well as smaller issues with the software.
