### 2020-10-20 04ef4a9
Update README (revise Intro, include Tutorial)
