+++ README.md
@@ -21,6 +21,7 @@
   - [Optional Arguments](#optional-arguments-1)
   - [Example](#example-1)
   - [Resulting Parameter Maps](#resulting-parameter-maps)
+- [Tutorial](#tutorial)
 - [Performance Metrics](#performance-metrics)
 - [Authors](#authors)
 - [References](#references)
@@ -33,9 +34,10 @@
 ## Introduction 
 
 *Scattered Light Imaging (SLI)* is a novel neuroimaging technique that allows to explore the substructure of nerve fibers, especially in regions with crossing nerve fibers, in whole brain sections with micrometer resolution ([Menzel et al. (2020)](https://arxiv.org/abs/2008.01037)). By illuminating histological brain sections from different angles and measuring the transmitted light under normal incidence, characteristic light intensity profiles (SLI profiles) can be obtained which provide crucial information such as the directions of crossing nerve fibers in each measured image pixel. 
-<!-- Details of the SLI measurement and the evaluation of the SLI profiles are described [below](#sli-measurement). -->
 
-This repository contains the *Scattered Light Imaging ToolboX (SLIX)* &ndash; an open-source Python package that allows a fully automated evaluation of SLI measurements and the generation of different parameter maps. With [`SLIXLineplotParameterGenerator`](#evaluation-of-sli-profiles), it is possible to evaluate individual SLI profiles and compute characteristics such as the number of peaks, their positions, and in-plane fiber direction angles. For a given SLI image stack, [`SLIXParameterGenerator`](#generation-of-parameter-maps) is able to compute up to 12 parameter maps, providing information about the measured brain tissue sample such as the individual in-plane direction angles of the nerve fibers for regions with up to three crossing nerve fiber bundles.
+This repository contains the *Scattered Light Imaging ToolboX (SLIX)* &ndash; an open-source Python package that allows a fully automated evaluation of SLI measurements and the generation of different parameter maps. The purpose of SLIX is twofold: 
+First, it allows to transform the raw data of SLI measurements (SLI image stack) to human-readable parameter maps that can be used for further analysis and interpreted by researchers. To this end, SLIX also contains functions to visualize the resulting parameter maps, e.g. as colored vector maps.
+Second, the results of SLIX can be easily processed further for use in tractography algorithms. For example, the resulting fiber direction maps can be stored as angles or as unit vectors, which can be used as input for streamline tractography algorithms ([Nolden et al. (2019)](https://doi.org/10.1007/978-3-658-25326-4_17)).
 
[…]
