### 2020-10-20 3cded17
Changed visualization example to also include the command line version of SLIX.
Changed the toolbox slightly to remove issues because lines were printed before tqdm was used.

Fixed some bugs in visualization.py that could result in a wrong orientation of the parameter maps.

Signed-off-by: Jan Andre Reuter <jan.andre.reuter@hotmail.de>
