### 2020-10-20 4da9f8d
Added documentation for visualization.py
Added first version of docs in repository which will be also added to the Wiki

Signed-off-by: Jan Andre Reuter <jan.andre.reuter@hotmail.de>
