+++ SLIX/visualization.py
@@ -19,8 +19,8 @@ def unit_vectors(directions):
 
     Returns
     -------
-    res : 3d-array, 3d-array, 3d-array
-        x-, y- and z-vector component in arrays
+    res : 3d-array, 3d-array
+        x- and y-vector component in arrays
     """
     directions_rad = numpy.deg2rad(directions)
     UnitX = -numpy.sin(0.5 * numpy.pi) * numpy.cos(directions_rad)
@@ -32,7 +32,25 @@ def unit_vectors(directions):
     return UnitX, UnitY
 
 
-def downsample(image, sample_size=10, background_value=-1, background_threshold=0.5):
+def downsample(image, sample_size, background_value=-1, background_threshold=0.5):
+    """
+    Reduce image dimensions of a parameter map by applying a median filter in each image in the z-axis.
+    The background will not be considered for the median filter except when the magnitude of it is above the given
+    threshold.
+
+    Parameters
+    ----------
+    image: 2D or 3D parameter map calculated with SLIX.toolbox.
+    sample_size: Down sampling parameter.
+    background_value: Background value in this parameter map. This is generally -1 but can differ for unit vectors.
+    background_threshold: If magnitude of the background values exceeds this value, the downsampled image will have
+                          background_value as it's resulting pixel value.
+
+    Returns
+    -------
+    2D or 3D Numpy array with reduced image dimensions
+    """
+    image = numpy.array(image)
     # downsample image
     if len(image.shape) == 2:
         x, y = image.shape
@@ -63,11 +81,30 @@ def downsample(image, sample_size=10, background_value=-1, background_threshold=
 
 def visualize_parameter_map(parameter_map, fig=None, ax=None, alpha=1,
                             cmap='viridis', vmin=0, vmax=None, colorbar=True):
+    """
+    This method will create a Matplotlib plot based on imshow to represent the given parameter map.
+    Here, the parameter map be plotted to the current axis and figure. If none is applied, the method will create a new
+    subfigure. To show the results, please use pyplot.show().
+
+    Parameters
+    ----------
+    parameter_map: 2D parameter map calculated with SLIX.toolbox.
+    fig: Matplotlib figure. If None a new subfigure will be created for fig and ax.
+    ax: Matplotlib axis. If None a new subfigure will be created for fig and ax.
+    alpha: Apply alpha to Matplotlib plots to overlay them with some other plots like the original measurement.
+    cmap: Matplotlib color map which is used for the shown image.
+    vmin: Minimum value in the resulting plot. If any value is below vmin it will be displayed in black.
+    vmax: Maximum value in the resulting plot. If any value is above vmax it will be displayed in white.
+    colorbar: Boolean value controlling if a color bar will be displayed in the current subplot.
+
+    Returns
+    -------
+    The current Matplotlib figure and axis. The image can be shown with pyplot.show().
+    """
     if fig is None or ax is None:
         fig, ax = plt.subplots(1, 1)
 
     cmap_mod = copy.copy(plt.get_cmap(cmap))
-    #parameter_map = numpy.swapaxes(parameter_map, -1, 0)
     im = ax.imshow(parameter_map, interpolation='nearest', cmap=cmap_mod, alpha=alpha)
     im.cmap.set_under(color='k')  # Color for values less than vmin
     im.cmap.set_over(color='w')  # Color for values more than vmax
@@ -78,7 +115,29 @@ def visualize_parameter_map(parameter_map, fig=None, ax=None, alpha=1,
     return fig, ax
 
 
-def visualize_unit_vectors(UnitX, UnitY, thinout=1, ax=None, alpha=0.8, background_threshold=0.5):
+def visualize_unit_vectors(UnitX, UnitY, thinout=1, ax=None, alpha=1, background_threshold=0.5):
+    """
[…]
