### 2020-10-20 ccab408
Fixed a bug with the unit maps where they could be overwritten by visualization methods
Signed-off-by: Jan Andre Reuter <jan.andre.reuter@hotmail.de>
