### 2020-10-20 be50eab
Update README (insert new vector map)
