### 2020-10-20 6cb206c
Update CONTRIBUTING.md
