### 2020-10-20 63a0f93
Fixed some typos in visualization.py
Fixed a bug in SLIXParameterGenerator where the description of some parameter would prevent the help option from functioning.

Reworked ipynb

Signed-off-by: Jan Andre Reuter <jan.andre.reuter@hotmail.de>
