### 2020-10-20 8d9e322
Minor changes in Doku (toolbox.py and visualization.py)
