+++ SLIX/toolbox.py
@@ -27,12 +27,12 @@ def all_peaks(line_profile, cut_edges=True):
 
     Parameters
     ----------
-    line_profile: 1D-NumPy array with all measurements of a single pixel.
-    cut_edges: When True only consider peaks within the second third of all detected peaks.
+    line_profile: 1D-NumPy array with all intensity values of a single image pixel in the stack.
+    cut_edges: If True, only consider peaks within the second third of all detected peaks.
 
     Returns
     -------
-    List with the positions of all detected peak positions.
+    List with the positions of all detected peaks.
     """
     number_of_measurements = line_profile.shape[0] // 2
 
@@ -61,7 +61,7 @@ def num_peaks_image(roiset, low_prominence=TARGET_PROMINENCE, high_prominence=nu
     roiset: Full SLI measurement (series of images) which is prepared for the pipeline using the SLIX toolbox methods.
     low_prominence: Lower prominence bound for detecting a peak.
     high_prominence: Higher prominence bound for detecting a peak.
-    cut_edges: When True only consider peaks within the second third of all detected peaks.
+    cut_edges: If True, only consider peaks within the second third of all detected peaks.
 
     Returns
     -------
@@ -117,7 +117,7 @@ def accurate_peak_positions(peak_positions, line_profile, low_prominence=TARGET_
 
     Returns
     -------
-    NumPy array with the positions of all detected peak positions.
+    NumPy array with the positions of all detected peaks.
     """
     n_roi = normalize(line_profile)
     peak_prominence = numpy.array(peak_prominences(n_roi, peak_positions)[0])
@@ -136,7 +136,7 @@ def peakdistance(peak_positions, number_of_measurements):
     Parameters
     ----------
     peak_positions: Detected peak positions of the 'all_peaks' method.
-    number_of_measurements: Number of measurements during a full SLI measurement, i.e. the number of points in one
+    number_of_measurements: Number of images in the SLI image stack, i.e. the number of points in one
     line profile.
 
     Returns
@@ -173,7 +173,7 @@ def peakdistance_image(roiset, low_prominence=TARGET_PROMINENCE, high_prominence
     roiset: Full SLI measurement (series of images) which is prepared for the pipeline using the SLIX toolbox methods.
     low_prominence: Lower prominence bound for detecting a peak.
     high_prominence: Higher prominence bound for detecting a peak.
-    cut_edges: When True only consider peaks within the second third of all detected peaks.
+    cut_edges: If True, only consider peaks within the second third of all detected peaks.
     centroid_calculation: Use centroid calculation to better determine the peak position regardless of the number of
     measurements / illumination angles used.
 
@@ -247,7 +247,7 @@ def prominence_image(roiset, low_prominence=TARGET_PROMINENCE, high_prominence=n
     roiset: Full SLI measurement (series of images) which is prepared for the pipeline using the SLIX toolbox methods.
     low_prominence: Lower prominence bound for detecting a peak.
     high_prominence: Higher prominence bound for detecting a peak.
-    cut_edges: When True only consider peaks within the second third of all detected peaks.
+    cut_edges: If True, only consider peaks within the second third of all detected peaks.
 
     Returns
     -------
@@ -318,7 +318,7 @@ def peakwidth_image(roiset, low_prominence=TARGET_PROMINENCE, high_prominence=nu
     roiset: Full SLI measurement (series of images) which is prepared for the pipeline using the SLIX toolbox methods.
     low_prominence: Lower prominence bound for detecting a peak.
     high_prominence: Higher prominence bound for detecting a peak.
-    cut_edges: When True only consider peaks within the second third of all detected peaks.
[…]
