+++ SLIX/visualization.py
@@ -10,16 +10,16 @@ CPU_COUNT = toolbox.CPU_COUNT
 
 def unit_vectors(directions):
     """
-    Calculates the unit vector from direction and inclination
+    Calculate the unit vectors (UnitX, UnitY) from a given direction angle.
 
     Parameters
     ----------
-    directions : 3d-array
-        direction in radian
+    directions: 3D NumPy array
+        direction angles in degrees
 
     Returns
     -------
-    res : 3d-array, 3d-array
+    UnitX, UnitY: 3D NumPy array, 3D NumPy array
         x- and y-vector component in arrays
     """
     directions_rad = numpy.deg2rad(directions)
@@ -32,23 +32,22 @@ def unit_vectors(directions):
     return UnitX, UnitY
 
 
-def downsample(image, sample_size, background_value=-1, background_threshold=0.5):
+def downsample(image, kernel_size, background_value=-1, background_threshold=0.5):
     """
-    Reduce image dimensions of a parameter map by applying a median filter in each image in the z-axis.
-    The background will not be considered for the median filter except when the magnitude of it is above the given
+    Reduce image dimensions of a parameter map by replacing (N x N) pixels by their median value for each image.
+    Image pixels with undefined values (background) will not be considered for computing the median, except when the fraction of background pixels lies above the defined
     threshold.
 
     Parameters
     ----------
-    image: 2D or 3D parameter map calculated with SLIX.toolbox.
-    sample_size: Down sampling parameter.
-    background_value: Background value in this parameter map. This is generally -1 but can differ for unit vectors.
-    background_threshold: If magnitude of the background values exceeds this value, the downsampled image will have
-                          background_value as it's resulting pixel value.
-
+    image: 2D or 3D parameter map (single image or image stack) calculated with SLIX.toolbox.
+    kernel_size: Downsampling parameter N (defines how many image pixels (N x N) are replaced by their median value).
+    background_value: Background value of the parameter map. This is generally -1 but can differ for unit vector maps.
+    background_threshold: Fraction of background pixels in the considered (N x N) area for which the image pixels are set to background_value. If the fraction of background pixels lies below this defined threshold, background pixels will not be considered for computing the median.
+    
     Returns
     -------
-    2D or 3D Numpy array with reduced image dimensions
+    2D or 3D NumPy array with reduced image dimensions.
     """
     image = numpy.array(image)
     # downsample image
@@ -58,16 +57,16 @@ def downsample(image, sample_size, background_value=-1, background_threshold=0.5
     else:
         x, y, z = image.shape
 
-    nx = numpy.ceil(x / sample_size).astype('int')
-    ny = numpy.ceil(y / sample_size).astype('int')
+    nx = numpy.ceil(x / kernel_size).astype('int')
+    ny = numpy.ceil(y / kernel_size).astype('int')
     small_img = pymp.shared.array((nx, ny, z), dtype='float32')
 
     for sub_image in range(z):
         with pymp.Parallel(CPU_COUNT) as p:
             for i in p.range(0, nx):
                 for j in range(0, ny):
-                    roi = image[sample_size * i:sample_size * i + sample_size,
-                                sample_size * j:sample_size * j + sample_size, sub_image]
+                    roi = image[kernel_size * i:kernel_size * i + kernel_size,
+                                kernel_size * j:kernel_size * j + kernel_size, sub_image]
                     if numpy.count_nonzero(roi == background_value) < background_threshold * roi.size:
                         small_img[i, j, sub_image] = numpy.median(roi[roi != background_value])
                     else:
@@ -82,19 +81,19 @@ def downsample(image, sample_size, background_value=-1, background_threshold=0.5
[…]
