+++ CONTRIBUTING.md
@@ -1,6 +1,6 @@
 # Contributing to SLIX (Scattered Light Imaging ToolboX)
 
-We would love your input to this repository! We want to make contributing to this project as easy and transparent as possible, whether it's:
+We would love your input to this repository! We want to make contributing to this project as easy and transparent as possible, whether it is:
 
 - Reporting a bug
 - Discussing the current state of the code
@@ -9,12 +9,12 @@ We would love your input to this repository! We want to make contributing to thi
 
 The development of SLIX is done through GitHub. A copy of the repository is also hosted on the [Forschungszentrum Jülich GitLab](https://jugit.fz-juelich.de/j.reuter/slix). All images shown in the [README.md](https://github.com/3d-pli/SLIX/blob/master/README.md) are hosted there.
 
-Pull requests, issues and feature requests are accepted via GitHub. If you plan to contribute to SLIX please follow the guidelines below.
+Pull requests, issues, and feature requests are accepted via GitHub. If you plan to contribute to SLIX, please follow the guidelines below.
 
 ## Seek support
-The Scattered Light Imaging ToolboX is maintained by Jan André Reuter and Miriam Menzel. For bug reports, feature requests and pull requests, please read the instructions below. For further support, you can contact both per mail.
+The Scattered Light Imaging ToolboX is maintained by Jan André Reuter and Miriam Menzel. For bug reports, feature requests, and pull requests, please read the instructions below. For further support, you can contact both per e-mail.
 
-| Person           | Mail adress            |
+| Person           | E-mail address         |
 | ---------------- | ---------------------- |
 | Jan André Reuter | j.reuter@fz-juelich.de |
 |    Miriam Menzel | m.menzel@fz-juelich.de |
@@ -25,22 +25,23 @@ The Scattered Light Imaging ToolboX is maintained by Jan André Reuter and Miria
 We use GitHub issues to track public bugs. Report a bug by opening a new issue [here](https://github.com/3d-pli/SLIX/issues).
 Write bug reports with detail, background, and add sample code if possible.
 
-Great Bug Reports tend to have:
+A good bug report should contain:
 
 - A quick summary and/or background
 - Steps to reproduce
     - Be specific!
-    - Give sample code if you can. 
+    - Give sample code if you can
 - What you expected would happen
 - What actually happens
-- Notes (possibly including why you think this might be happening, or stuff you tried that didn't work)
+- Notes (possibly including why you think this might be happening, or stuff you tried that did not work)
 
 ## Feature additions
-SLIX is a project to support users in the evaluation of scatter measurements of brain slices. For this purpose, several parameter maps are already generated, which can be viewed for further analysis. A small visualization toolbox is also provided.
 
-While these features are sufficient for the current project framework, other users might have additional requirements for the software. Therefore pull-requests with further features and the reporting of errors in the program are welcome.
+SLIX is a project to support users in the evaluation of Scattered Light Imaging (SLI) measurements of brain slices. For this purpose, several parameter maps are generated with the toolbox. For further analysis, the parameter maps can be visualized using an additional visualization toolbox.
 
-The toolbox currently has two main files which contain most of the source code: `toolbox.py` and `visualization.py`. The first file does most of the work containing methods for preparing the measurement, generating the parameter arrays and reshaping them to an image. Visualization contains all methods to show generated parameter maps or even convert the direction maps to unit vectors. If any new features shall be added, consider if a new file is more fitting than implementing the new functions into an already existing file.
+While these features are sufficient for the current framework of the project, other users might have additional requirements for the software. Therefore, pull-requests including additional features and the reporting of errors in the program are welcome.
+
[…]
