+++ README.md
@@ -267,7 +267,7 @@ The in-plane direction angles are only computed if the SLI profile has one, two,
 To obtain a measure for the signal-to-noise, the difference between maximum and minimum can be divided by the average.
 
 ## Tutorial
-The [Jupyter notebook](https://github.com/3d-pli/SLIX/blob/3-vector-visualization/examples/Visualization_Example.ipynb) demonstrates how SLIX can be used to analyze SLI measurements and to visualize the results. For example, it allows to display the generated parameter maps in different colors, and to show the orientations of (crossing) nerve fibers as colored lines (vector maps) by computing unit vector maps from the direction maps. The following vector map has been generated with the function `visualize_unit_vectors`, using `alpha = 0.8` (defining the transparency of the background image), `thinout = 30` (i.e. 30 x 30 pixels were evaluated together), and `background_threshold = 0.7` (i.e. if more than 70% of the evaluated pixels are `-1`, no vector will be computed). 
+The [Jupyter notebook](https://github.com/3d-pli/SLIX/blob/master/examples/Visualization_Example.ipynb) demonstrates how SLIX can be used to analyze SLI measurements and to visualize the results. For example, it allows to display the generated parameter maps in different colors, and to show the orientations of (crossing) nerve fibers as colored lines (vector maps) by computing unit vector maps from the direction maps. The following vector map has been generated with the function `visualize_unit_vectors`, using `alpha = 0.8` (defining the transparency of the background image), `thinout = 30` (i.e. 30 x 30 pixels were evaluated together), and `background_threshold = 0.7` (i.e. if more than 70% of the evaluated pixels are `-1`, no vector will be computed). 
 
 <img src="https://jugit.fz-juelich.de/j.reuter/slix/-/raw/assets/output_unit_vectors.png" height="327">
 
