+++ doc/SLIX/toolbox.md
@@ -22,7 +22,7 @@ Functions
     
     Returns
     -------
-    NumPy array with the positions of all detected peak positions.
+    NumPy array with the positions of all detected peaks.
 
     
 `all_peaks(line_profile, cut_edges=True)`
@@ -31,12 +31,12 @@ Functions
     
     Parameters
     ----------
-    line_profile: 1D-NumPy array with all measurements of a single pixel.
-    cut_edges: When True only consider peaks within the second third of all detected peaks.
+    line_profile: 1D-NumPy array with all intensity values of a single image pixel in the stack.
+    cut_edges: If True, only consider peaks within the second third of all detected peaks.
     
     Returns
     -------
-    List with the positions of all detected peak positions.
+    List with the positions of all detected peaks.
 
     
 `centroid_correction(line_profile, peak_positions, low_prominence=0.08, high_prominence=inf)`
@@ -58,8 +58,7 @@ Functions
 
     
 `create_background_mask(IMAGE, threshold=10)`
-:   Creates a background mask based on given threshold. As all background pixels are near zero when looking through
-    the z-axis plot, this method should remove most of the background allowing for better approximations using the
+:   Creates a background mask by setting all image pixels with low scattering signals to zero. As all background pixels are near zero for all images in the SLI image stack, this method should remove most of the background allowing for better approximations using the
     available features. It is advised to use this function.
     
     Arguments:
@@ -73,7 +72,7 @@ Functions
 
     
 `create_roiset(IMAGE, ROISIZE, extend=True)`
-:   Create roi set of given image by creating an image containing the average value of pixels within the
+:   Create roi set of the given image by creating an image containing the average value of pixels within the
     specified ROISIZE. The returned image will have twice the size in the third axis as the both halfs will be doubled
     for the peak detection.
     
@@ -92,10 +91,10 @@ Functions
     line_profile: Original line profile used to detect all peaks. This array will be further
     analyzed to better determine the peak positions.
     peak_positions: Detected peak positions of the 'all_peaks' method.
-    left_bound: Left bound for linear interpolation
-    right_bound: Right bound for linear interpolation
-    target_peak_height: Targeted peak height for centroid calculation
-    number_of_samples: Number of samples used for linear interpolation
+    left_bound: Left bound for linear interpolation.
+    right_bound: Right bound for linear interpolation.
+    target_peak_height: Targeted peak height for centroid calculation.
+    number_of_samples: Number of samples used for linear interpolation.
     
     Returns
     -------
@@ -135,7 +134,7 @@ Functions
     roiset: Full SLI measurement (image series) which is prepared for the pipeline using the SLIX toolbox methods.
     low_prominence: Lower prominence bound for detecting a peak.
     high_prominence: Higher prominence bound for detecting a peak.
-    cut_edges: When True only consider peaks within the second third of all detected peaks.
+    cut_edges: If True, only consider peaks within the second third of all detected peaks.
     
     Returns
     -------
@@ -147,12 +146,12 @@ Functions
 `non_crossing_direction(peak_positions, number_of_measurements)`
 :   Calculate one direction angle based on the given peak positions. If more than two peaks are present, no
     direction angle will be calculated to avoid errors. This will result in a direction angle of BACKGROUND_COLOR.
-    The peak positions are determined by the position of the corresponding peak pair.
+    The direction angle is determined by the mid position between two peaks.
     
     Parameters
     ----------
     peak_positions: Detected peak positions of the 'all_peaks' method.
-    number_of_measurements: Number of measurements during a full SLI measurement, i.e. the number of points in the line
[…]
