### 2020-10-20 3d-pli/SLIX#8 @alexrockhill
The documentation looks good, the only issue I'd be concerned with is missing or corrupted data. What if one of the images is not to be included because it has been corrupted?
