### 2020-10-20 3d-pli/SLIX#7 @miriammenzel
We have added the automatic generation of unit vectors from a given direction image (visualization.py). The unit vectors could be used for example as input for streamline tractography algorithms. The [tutorial](https://github.com/3d-pli/SLIX/blob/master/examples/Visualization_Example.ipynb) shows how the vector map can be visualized (similar to the vectors shown in figure 2i, see last image in the tutorial) so that the user is now able to reproduce all figures, including the vector maps. Furthermore, we added a statement to the [Introduction](https://github.com/3d-pli/SLIX#introduction) to explain the purpose of SLIX: it allows to transform the raw data of SLI measurements to parameter maps that can be interpreted by researchers or be used for further analysis, e.g. tractography. If other output formats are required, this can be easily added to the software. To explain the user how to contribute to SLIX, e.g. adding new features, we added the section “Feature additions” in the [CONTRIBUTING.md](https://github.com/3d-pli/SLIX/blob/master/CONTRIBUTING.md).
@alexrockhill @matteomancini: Please let us know if this now meets your requirements and if we can close this issue.
