### 2020-10-20 3d-pli/SLIX#6 @miriammenzel
In addition to what I already mentioned, we included the option `--with_mask` in the [tutorial](https://github.com/3d-pli/SLIX/blob/master/examples/Visualization_Example.ipynb) to demonstrate the user how to use it and what the resulting parameter maps look like.

@alexrockhill: Please let us know if this meets your requirements and if we can close this issue.
