### 2020-10-20 3d-pli/SLIX#4 @miriammenzel
We have now added a [tutorial](https://github.com/3d-pli/SLIX/blob/master/examples/Visualization_Example.ipynb) (Jupyter notebook) that walks the user step by step through the computation of the parameter maps and shows how the parameter maps can be visualized. We have added the file `visualization.py` that allows to display parameter maps with a certain color map (using matplotlib), and to visualize the fiber orientations as vector maps (colored lines displayed on top of the original image), which is also demonstrated in the tutorial. In addition, we included everything in a Wiki and also added API documentation for toolbox.py and visualization.py.
@alexrockhill @matteomancini: Please let us know if this meets your requirements and if we can close this issue.
