### 2020-10-20 dda166d
Added error checks for file format and shape
Signed-off-by: Jan Andre Reuter <j.reuter@fz-juelich.de>
