### 2020-10-20 3d-pli/SLIX#7 @alexrockhill
The tutorial looks really good, and is much more what I would consider as meeting the requirements for a full solution. I'll take a look later about the actual results now that the process can be worked through in python to the unit vector visualization.

The title of this issue is on the integration with tractography software, if there were to be one or two cells added at the end of the tutorial on how to integrate with your preferred tractography software I still think that would be a helpful addition.
