### 2020-10-20 3d-pli/SLIX#8 @miriammenzel
We expect that it is extremely unlikely that a whole image is corrupted/missing if the SLI measurement was performed in the suggested way (and the user should notice this immediately after the measurement). Moreover, the loss of information is quite large (e.g. missing peaks in the SLI profile could lead to different results). Therefore, we suggest to repeat the measurement instead of correcting this in SLIX.  If single image pixels have been corrupted, the tool automatically copes with that. We will add a specific error message if the input format is not as required.
