### 2020-10-21 e2d8d52
Fixed minor documentation bug in README.md
