### 2020-10-21 7c5f7ce
Changed distribution type from setup.py to pep517
Expand workflow.

Signed-off-by: Jan Andre Reuter <jan.andre.reuter@hotmail.de>
