+++ setup.cfg
@@ -1,0 +1,35 @@
+[metadata]
+name = SLIX
+version = v1.2.1
+author = Jan Andre Reuter, Miriam Menzel
+author_email = j.reuter@fz-juelich.de, m.menzel@fz-juelich.de
+description = SLIX allows an automated evaluation of SLI measurements and generates different parameter maps.
+license_file = LICENSE
+long_description = file: README.md
+license = MIT
+classifiers =
+    Programming Language :: Python :: 3
+    License :: OSI Approved :: MIT License
+    Operating System :: OS Independent
+
+[options]
+zip_safe = False
+include_package_data = True
+packages = find:
+scripts =
+    bin/SLIXParameterGenerator
+    bin/SLIXLineplotParameterGenerator
+install_requires =
+    numpy
+    scipy
+    matplotlib
+    pymp-pypi
+    python_version >= "3.6"
+    nibabel
+    tifffile
+    tqdm
+    Pillow
+tests_require =
+    pytest
+    flake8
+    pytest-cov
