+++ .github/workflows/publish-to-pypi.yml
@@ -1,0 +1,39 @@
+name: Publish Python 🐍 distributions 📦 to PyPI and TestPyPI
+
+on: [push]
+
+jobs:
+  runs-on: ubuntu-latest
+  strategy:
+    matrix:
+      python-version: [ 3.6, 3.7, 3.8, 3.9 ]
+
+  steps:
+    - uses: actions/checkout@v2
+
+    - name: Set up Python ${{ matrix.python-version }}
+      uses: actions/setup-python@v2
+      with:
+        python-version: ${{ matrix.python-version }}
+
+    - name: Install dependencies
+      run: |
+        python -m pip install --upgrade pip
+        pip install setuptools wheel
+        if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
+
+    - name: Build
+      run: |
+        python setup.py sdist bdist_wheel
+
+    - name: Publish distribution 📦 to Test PyPI
+        uses: pypa/gh-action-pypi-publish@master
+          with:
+            password: ${{ secrets.test_pypi_password }}
+            repository_url: https://test.pypi.org/legacy/
+
+    - name: Publish distribution 📦 to PyPI
+        if: startsWith(github.ref, 'refs/tags')
+          uses: pypa/gh-action-pypi-publish@master
+            with:
+              password: ${{ secrets.pypi_password }}
