### 2020-10-21 f5445e0
Update to workflow
Signed-off-by: Jan Andre Reuter <jan.andre.reuter@hotmail.de>
