+++ .github/workflows/publish-to-pypi.yml
@@ -1,4 +1,4 @@
-name: Publish Python 🐍 distributions 📦 to PyPI and TestPyPI
+name: Publish package
 
 on: [push]
 
@@ -6,7 +6,7 @@ jobs:
   runs-on: ubuntu-latest
   strategy:
     matrix:
-      python-version: [ 3.6, 3.7, 3.8, 3.9 ]
+      python-version: [ 3.6, 3.7, 3.8 ]
 
   steps:
     - uses: actions/checkout@v2
@@ -22,18 +22,31 @@ jobs:
         pip install setuptools wheel
         if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
 
-    - name: Build
+    - name: Install pep517
       run: |
-        python setup.py sdist bdist_wheel
+        python -m
+        pip install
+        pep517
+        --user
 
-    - name: Publish distribution 📦 to Test PyPI
-        uses: pypa/gh-action-pypi-publish@master
-          with:
-            password: ${{ secrets.test_pypi_password }}
-            repository_url: https://test.pypi.org/legacy/
+    - name: Build a binary wheel and a source tarball
+      run: |
+        python -m
+        pep517.build
+        --source
+        --binary
+        --out-dir dist/
 
-    - name: Publish distribution 📦 to PyPI
-        if: startsWith(github.ref, 'refs/tags')
-          uses: pypa/gh-action-pypi-publish@master
-            with:
-              password: ${{ secrets.pypi_password }}
+    - name: Publish package to TestPyPI
+      uses: pypa/gh-action-pypi-publish@master
+      with:
+        user: __token__
+        password: ${{ secrets.test_pypi_password }}
+        repository_url: https://test.pypi.org/legacy/
+
+    - name: Publish package
+      if: github.event_name == 'push' && startsWith(github.ref, 'refs/tags')
+      uses: pypa/gh-action-pypi-publish@master
+      with:
+        user: __token__
+        password: ${{ secrets.pypi_password }}
