+++ setup.cfg
@@ -1,11 +1,12 @@
 [metadata]
 name = SLIX
-version = v1.2.1
+version = v1.2.1-1
 author = Jan Andre Reuter, Miriam Menzel
 author_email = j.reuter@fz-juelich.de, m.menzel@fz-juelich.de
 description = SLIX allows an automated evaluation of SLI measurements and generates different parameter maps.
 license_file = LICENSE
 long_description = file: README.md
+long_description_content_type = text/markdown
 license = MIT
 classifiers =
     Programming Language :: Python :: 3
