### 2020-10-21 4318887
Fixed install errors because of faulty setup.cfg
