+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = SLIX
-version = v1.2.1-1
+version = v1.2.1-2
 author = Jan Andre Reuter, Miriam Menzel
 author_email = j.reuter@fz-juelich.de, m.menzel@fz-juelich.de
 description = SLIX allows an automated evaluation of SLI measurements and generates different parameter maps.
@@ -25,7 +25,6 @@ install_requires =
     scipy
     matplotlib
     pymp-pypi
-    python_version >= "3.6"
     nibabel
     tifffile
     tqdm
@@ -34,3 +33,5 @@ tests_require =
     pytest
     flake8
     pytest-cov
+python_requires =
+    >= 3.6
