### 2020-10-21 6b6a720
Changed version number. Fixed last problems with the packaging.
Signed-off-by: Jan Andre Reuter <jan.andre.reuter@hotmail.de>
