+++ .github/workflows/publish-to-pypi.yml
@@ -1,6 +1,8 @@
 name: Publish package
 
-on: [push]
+on:
+  push:
+    branches: [ master ]
 
 jobs:
   build:
