### 2020-10-21 3d-pli/SLIX#7 @miriammenzel
Thank you @alexrockhill for the suggestion. We have now added the section “Possible Integration in Tractograpy Algorithm” at the end of our [tutorial](https://github.com/3d-pli/SLIX/blob/master/examples/Visualization_Example.ipynb) to explain how the unit vector maps obtained by SLIX can be prepared to be used as input in streamline-based tractography algorithms (seed points propagating through the vector field). As we currently have no data set available that could be used as input (as mentioned above, we would need SLI measurements of several consecutive brain sections to generate a sufficient 3D volume of unit vectors), we show an example where a 3D vector field (obtained by 3D Polarized Light Imaging) has been used as input for in-house developed tractography software.
@alexrockhill: Please let us know if this meets your requirements and if we can close this issue.
