### 2020-10-21 f022c0e
Automatic PyPi publishing (#15)
* Changed distribution type from setup.py to pep517

* Update to workflow

* Changed version number. Fixed last problems with the packaging.

* Update README.md

Changed install instructions to represent new way of installation
