+++ SLIX/__init__.py
@@ -1,0 +1,5 @@
+"""Module docstring text"""
+__version__ = 'v1.2.1'
+
+from . import toolbox
+from . import visualization
