### 2020-10-21 #2675 @miriammenzel
@alexrockhill: Thank you very much for your comments and sharing your thoughts. The purpose of SLIX is indeed to process the raw data from Scattered Light Imaging (SLI) and generate human-readable parameter maps. In some way, our software fulfills a similar purpose as software that transforms RF signals measured in an MR scanner to human-readable MRI images. As SLI is a completely new technique, SLIX is the first software of this kind and allows for the first time to evaluate the SLI measurements in a fully automated way and to study and interpret the scattering signals for the whole measured tissue sample. The SLI raw data is very complex and not easily interpretable. The purpose of SLIX is to process the raw data further, e.g. to extract the in-plane (crossing) nerve fiber orientations from the measured signals (peaks in the SLI profiles), and to provide interpretable parameter maps that can easily be processed/analyzed further.

We agree that the output of SLIX should be easily adaptable by the user to account for his/her diverse needs. Therefore, we provided “visualization.py” which contains different methods to visualize the output as described in our [tutorial](https://github.com/3d-pli/SLIX/blob/master/examples/Visualization_Example.ipynb). The API is flexible and the output can be easily changed by the user, e.g. by replacing the function `visualize_unit_vectors` by something else. We added a Wiki, a full documentation of all available functions, and a brief statement concerning “Feature additions” in CONTRIBUTING.md.
Concerning the input, it is already possible to perform the SLI measurement with arbitrary illumination angles and patterns, as long as the sample is illuminated under a fixed polar angle and different azimuthal angles in equidistant steps from 0° to 360°. These requirements are needed to perform a correct analysis of the SLI image stacks as described in the paper and should not be changed by the user.

We added a thorough documentation of why each analysis decision was taken in the software, both in the README.md and in the tutorial, and added options to change parameters (such as the 8% prominence, the 6% peak tip height, or the fraction of background pixels).
In this way, the user can test the parameters and understand what they are used for. In the tutorial, we walk the user through step-by-step from the raw data (SLI image stack) to the results (parameter maps), and show different possibilities of visualization.

To demonstrate how SLIX fits in a workflow and could be used by tractography software, we added the section "Possible Integration in Tractography Algorithm" at the end of the tutorial, as suggested. There, we explain how the unit vector maps obtained by SLIX can be prepared to be used as input in streamline-based tractography algorithms (seed points propagating through the vector field). As we currently have no data set available that could be used as input (we would need SLI measurements of several consecutive brain sections to generate a sufficient 3D volume of unit vectors), we show an example where a 3D vector field (obtained by 3D Polarized Light Imaging) has been used as input for in-house developed tractography software.

We hope that we have now addressed all issues raised by the reviewers so that our software paper can be published. We have replied to all open issues and ask the reviewers to close them if our responses have been sufficient.
@alexrockhill @matteomancini @rly: Please let us know if we should add anything else.
