### 2020-10-21 #2675 @rly
I am reviewing the most recent version of the SLIX GitHub repo, which seems to include quite a few improvements since the last review: the addition of an example jupyter notebook, edits to the readme and paper, bug fixes, improved documentation, and release on PyPI.

Installation of SLIX was easy, both using PyPI and installing from source. Thank you for adding the option to install via pip. The software is an important tool for processing data from a new imaging technique - scattered light imaging. The paper is concise and well written, the examples run well, and the documentation is comprehensive. As mentioned by @matteomancini , the comparison to other tools and state of the field is lacking in the paper. I think it would be valuable to state clearly that to your knowledge, there is no other tool that does what SLIX does and that it is a necessary step for processing scattered light imaging data.

Is there source code for the command line executables in `bin/`? It should also be noted in the README that these executables cannot be run on Windows (at least I could not get them to run on Windows and used Linux instead).

Minor comments:
- The Jupyter notebook has spelling errors, "paramter" and "scatterig", in a few places.
- Paper figure 1 caption: The cyan line should be labeled. "Peaks with lower prominences" -> "Peaks with lower prominences (cyan line)"
- Paper figure 1 caption: "The fiber direction angles φ1 and φ2 are computed from the mid positions of the prominent peak pairs." It is not clear here how the pairs are chosen.
- It might be useful to add a troubleshooting section in the documentation. e.g., if the algorithm finds 3 or 5 prominent peaks for a large chunk of an image, then it cannot compute a direction angle for those pixels. What should the user do to improve the algorithm? Should the parameters be tweaked to get a better estimate for those pixels, and if so, how should they be tweaked? e.g. increase the prominence threshold? Or is it not possible to salvage these sections of the image (or not worth the effort)?
