### 2020-10-21 #2675 @Thyre
Thanks so much for your comment.
Both programs in `/bin` are actually Python files itself. You should be able to view the source code in the repository without any problems. I also agree that we should add a note that the program cannot be executed on Windows for now. This is unfortunately caused by `pymp-pypi` and will be fixed soon when the implementation of the GPU variant is finished. Until then, only users using macOS and Linux are able to run the program. A workaround on Windows is to use the WSL in the meantime.

I absolutely agree on all of your minor comments. We will fix them as soon as possible.
