### 2020-10-21 5610b23
add statement to paper
