### 2020-10-21 #2675 @miriammenzel
@rly: Thank you very much for your comments and your positive feedback.
I’ve added a note in [README.md](https://github.com/3d-pli/SLIX/blob/master/README.md#installation-of-slix) that SLIX can currently only be executed under macOS and Linux.
I’ve also corrected the typos in the Jupyter notebook. (Thank you for the careful reading!).
In the [paper](https://github.com/3d-pli/SLIX/blob/master/paper/paper.pdf), I’ve added a clear statement at the end of the second paragraph that SLIX is the first software of its kind, and I also corrected the wording in the caption of figure 1 as suggested.
Finally, I’ve added a second paragraph to the section ["Direction Angles"](https://github.com/3d-pli/SLIX/blob/master/README.md#direction-angles) in the README, explaining what the user can do when three or five peaks are detected.
Please let us know if this meets your requirements of if there is anything else that we should add.
