### 2020-10-22 d0735d5
Fixed both CPU and GPU peakwidth
Signed-off-by: Jan Andre Reuter <jan.andre.reuter@hotmail.de>
