### 2020-10-22 3d-pli/SLIX#7 @matteomancini
I think that the notebook looks amazing and it really makes things clear!
