### 2020-10-22 3d-pli/SLIX#4 @matteomancini
As mentioned in #7 I think the notebook really does the job, also in terms of providing further examples!
