### 2020-10-22 #2675 @matteomancini
I think that both the tool and the paper now look much more complete. @Thyre I am glad to hear that the PyPI suggestion was useful! @miriammenzel I did not realise there was such a strict limit in terms of word count - taking that into account I think that the sentences added are sufficient to explain the difference between PLI and SLI, and the fact that SLIX is the first of its kind.
All my comments have been addressed, and once the last minor things noticed by @rly are fixed, I think that the tool and the paper are both suitable for publication on JOSS.
