### 2020-10-22 #2675 @rly
@Thyre @miriammenzel Thanks for addressing my concerns! Everything looks good on my end. I think that the tool and the paper are both suitable for publication on JOSS.
