+++ SLIX/toolbox.py
@@ -1,7 +1,3 @@
-import tifffile
-import nibabel
-import numpy
-
 try:
     try:
         import cupy
@@ -12,29 +8,6 @@ try:
 except ModuleNotFoundError:
     pass
 from SLIX.SLIX_CPU import toolbox as cpu_toolbox
-
-
-def read_image(filepath):
-    """
-    Reads image file and returns it.
-    Supported file formats: NIfTI, Tiff.
-
-    Arguments:
-        filepath: Path to image
-
-    Returns:
-        numpy.array: Image with shape [x, y, z] where [x, y] is the size of a single image and z specifies the number
-                     of measurements
-    """
-    # Load NIfTI dataset
-    if filepath.endswith('.nii'):
-        data = nibabel.load(filepath).get_fdata()
-        data = numpy.squeeze(numpy.swapaxes(data, 0, 1))
-    else:
-        data = tifffile.imread(filepath)
-        data = numpy.squeeze(numpy.moveaxis(data, 0, -1))
-
-    return data
 
 
 def peaks(image, use_gpu=True, return_numpy=True):
@@ -99,3 +72,7 @@ def centroid_correction(image, peak_image, low_prominence=cpu_toolbox.TARGET_PRO
         return gpu_toolbox.centroid_correction(image, peak_image, low_prominence, high_prominence, return_numpy)
     else:
         return cpu_toolbox.centroid_correction(image, peak_image, low_prominence, high_prominence)
+
+
+def unit_vectors(direction, use_gpu=True, return_numpy=True):
+    pass
