+++ SLIX/__init__.py
@@ -1,1 +1,3 @@
 from . import toolbox
+from . import io
+from . import visualization
