### 2020-10-23 d667e63
Implemented unit vector calculation for CPU and GPU. Tests aren't implemented yet.
Signed-off-by: Jan Andre Reuter <j.reuter@fz-juelich.de>
