+++ SLIX/__init__.py
@@ -1,5 +1,6 @@
 """Module docstring text"""
 __version__ = 'v2.0'
+
 from . import toolbox
 from . import io
 from . import visualization
