+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = SLIX
-version = v1.2.1-3
+version = v2.0
 author = Jan Andre Reuter, Miriam Menzel
 author_email = j.reuter@fz-juelich.de, m.menzel@fz-juelich.de
 description = SLIX allows an automated evaluation of SLI measurements and generates different parameter maps.
