### 2020-10-23 71c7e7f
Removed unit_vectors from visualization.py as it is now implemented in toolbox.py
Added imwrite method and hdf5 stubs to SLIX.io

Signed-off-by: Jan Andre Reuter <j.reuter@fz-juelich.de>
