+++ SLIX/io.py
@@ -1,9 +1,19 @@
 import numpy
 import tifffile
 import nibabel
+import h5py
+from PIL import Image
 
 
-def read_image(filepath):
+def hdf5_read(filepath, dataset):
+    pass
+
+
+def hdf5_write(filepath, dataset):
+    pass
+
+
+def imread(filepath):
     """
     Reads image file and returns it.
     Supported file formats: NIfTI, Tiff.
@@ -19,8 +29,18 @@ def read_image(filepath):
     if filepath.endswith('.nii'):
         data = nibabel.load(filepath).get_fdata()
         data = numpy.squeeze(numpy.swapaxes(data, 0, 1))
-    else:
+    elif filepath.endswith('.tiff') or filepath.endswith('.tif'):
         data = tifffile.imread(filepath)
         data = numpy.squeeze(numpy.moveaxis(data, 0, -1))
-
+    else:
+        data = numpy.array(Image.open(filepath))
     return data
+
+
+def imwrite(data, filepath):
+    if filepath.endswith('.nii'):
+        nibabel.save(nibabel.Nifti1Image(numpy.swapaxes(data, 0, 1), numpy.eye(4)), filepath)
+    elif filepath.endswith('.tiff') or filepath.endswith('.tif'):
+        tifffile.imwrite(filepath, numpy.moveaxis(data, -1, 0))
+    else:
+        Image.fromarray(data).save(filepath)
