### 2020-10-23 92196be
Added basic HDF5 functionality. Changed main.py to use io.py
Signed-off-by: Jan Andre Reuter <j.reuter@fz-juelich.de>
