### 2020-10-23 0e0b241
Fixed reading images in tests.
Fixed a bug in num_peaks which caused tests to fail.

Signed-off-by: Jan Andre Reuter <j.reuter@fz-juelich.de>
