+++ tests/test_compare_cpu_gpu.py
@@ -1,7 +1,6 @@
 import SLIX
 import numpy
 
-from matplotlib import pyplot as plt
 import tifffile
 
 
