+++ setup.cfg
@@ -3,6 +3,7 @@ name = SLIX
 version = v2.0
 author = Jan Andre Reuter, Miriam Menzel
 author_email = j.reuter@fz-juelich.de, m.menzel@fz-juelich.de
+url = https://github.com/3d-pli/SLIX
 description = SLIX allows an automated evaluation of SLI measurements and generates different parameter maps.
 license_file = LICENSE
 long_description = file: README.md
@@ -19,16 +20,17 @@ include_package_data = True
 packages = find:
 scripts =
     bin/SLIXParameterGenerator
-    bin/SLIXLineplotParameterGenerator
+    #bin/SLIXLineplotParameterGenerator
 install_requires =
     numpy
     scipy
     matplotlib
-    pymp-pypi
     nibabel
     tifffile
     tqdm
     Pillow
+    h5py
+    numba
 tests_require =
     pytest
     flake8
