### 2020-10-24 efc1ec5
added DOIs to paper.bib
