+++ paper/paper.bib
@@ -15,7 +15,6 @@
   Year                     = {2020},
   Pages                    = {021002},
   Volume                   = {10},
-
   Doi                      = {10.1103/PhysRevX.10.021002},
   Issue                    = {2},
   Owner                    = {mmenzel},
@@ -29,7 +28,8 @@
   Year                     = {2011},
   Number                   = {2},
   Pages                    = {1091--1101},
-  Volume                   = {54}
+  Volume                   = {54},
+  Doi                      = {10.1016/j.neuroimage.2010.08.075}
 }
 
 @Article{MAxer2011_2,
@@ -39,6 +39,7 @@
   Year                     = {2011},
   Number                   = {34},
   Pages                    = {1--13},
-  Volume                   = {5}
+  Volume                   = {5},
+  Doi                      = {10.3389/fninf.2011.00034}
 }
 
