### 2020-10-24 #2675 @miriammenzel
@oliviaguest We were not aware that the repository on zenodo needs to have the same title as the paper. We thought it should have the same title as the software and the documentation. Is this a problem? (I'm not sure if we can change this without changing the DOI.)
We have added the missing DOIs to the paper.
