### 2020-10-24 #2675 @oliviaguest
@miriammenzel see: https://help.zenodo.org/

>I only want to change the title of my upload, do I still get a new DOI?
>
> No, as before you can continue to edit the metadata of your upload without creating a new version of a record. You should only create a new version if you want to update the files of your record.
