### 2020-10-24 #2675 @miriammenzel
@oliviaguest I have changed the title of the repository on zenodo accordingly. Everything should be all set now. Thanks again for the review.
