### 2020-10-25 #2675 @kyleniemeyer
Hi @miriammenzel, I'm the EIC on duty this week doing final checks before publishing. Could you address these issues?

- [ ] The article is missing a [Statement of Need section](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain), which we have begun requiring for articles. You can likely use text that is already present in the article.
- [ ] There are a few citations that appear to use the incorrect command, such as in the first paragraph, and end up with nested parentheses. I think you should use the `[@author:2001]` command for those instead. For reference, the options are:
```
- `@author:2001`  ->  "Author et al. (2001)"
- `[@author:2001]` -> "(Author et al., 2001)"
- `[@author1:2001; @author2:2001]` -> "(Author1 et al., 2001; Author2 et al., 2002)"
```

Thanks!
