+++ paper/paper.md
@@ -23,7 +23,7 @@ bibliography: paper.bib
 
 # Summary
 
-To study the structure and function of the brain, a detailed reconstruction of the brain's nerve fiber architecture is required. However, many neuroimaging techniques, such as 3D-Polarized Light Imaging (@MAxer2011_1. @MAxer2011_2) struggle with the correct reconstruction of crossing nerve fibers. 
+To study the structure and function of the brain, a detailed reconstruction of the brain's nerve fiber architecture is required. However, many neuroimaging techniques, such as 3D-Polarized Light Imaging [@MAxer2011_1; @MAxer2011_2] struggle with the correct reconstruction of crossing nerve fibers. 
 The recently developed technique *Scattered Light Imaging (SLI)* uses light in the optical range to explore the substructure of (crossing) nerve fibers in whole brain sections with micrometer resolution, providing crucial information such as the orientations of crossing nerve fibers in a measured image pixel [@menzel2020-NeuroImage]. The measurement principle was first introduced by @menzel2020. Unstained histological brain sections are illuminated from different angles and the transmitted (scattered) light behind the sample is measured under normal incidence (see Figure \ref{figure}a). Every pixel in the resulting SLI image series contains a light intensity profile (*SLI profile* $I(\phi)$, cf. Figure \ref{figure}b-c) which is characteristic for the brain tissue structure at this point. The number and position of significant peaks in the SLI profiles can be used, for example, to determine the individual orientations of (crossing) nerve fibers in each image pixel. 
 
 Here, we present the *Scattered Light Imaging ToolboX* (`SLIX`) &ndash; an open-source Python package that allows for the first time a fully automated evaluation of SLI measurements and the generation of different parameter maps (see Figure \ref{figure2}a-h) which reflect different characteristics of the SLI profiles: the average, the number of peaks, the average peak prominence, the average peak width, the distance between peaks, and the position of the peaks from which the in-plane direction angles of the nerve fibers are computed in regions with up to three crossing nerve fiber bundles. The user can choose which parameter maps are generated by using specific command line parameters. The resulting direction maps can be used to visualize the individual orientations of in-plane (crossing) nerve fibers, as shown in Figure \ref{figure2}i. The purpose of SLIX is to process the SLI raw data and to provide interpretable parameter maps that can easily be processed and analyzed by other researchers. As SLI is a completely new technique and there exist no comparable methods to our knowledge, SLIX is the first software of its kind and a necessary step for processing scattered light imaging data.
@@ -37,11 +37,12 @@ For a full documentation of `SLIX`, the reader is referred to our [GitHub page](
 
[…]
