### 2020-10-25 #2675 @miriammenzel
@kyleniemeyer: Thanks for letting me know. I have added a Statement of Need section at the end of our [paper](https://github.com/3d-pli/SLIX/blob/master/paper/paper.pdf) and corrected the citations in the first paragraph. Please let me know if everything is OK now.
