### 2020-10-25 #2675 @oliviaguest
@Thyre can you merge and compile the PDF again with this change, please? https://github.com/3d-pli/SLIX/pull/16
