+++ paper/paper.bib
@@ -10,7 +10,7 @@
 
 @Article{menzel2020,
   Title                    = {Toward a High-resolution Reconstruction of {3D} Nerve Fiber Architectures and Crossings in the Brain Using Light Scattering Measurements and Finite-Difference Time-Domain Simulations},
-  Author                   = {M.~Menzel and M.~Axer and H.~De~Raedt and I.~Costantini and L.~Silvestri and F.~S.~Pavone and K.~Amunts and K.~Michielsen},
+  Author                   = {Menzel, M. and Axer, M. and De~Raedt, H. and Costantini, I. and Silvestri, L. and Pavone, F.~S. and Amunts, K. and Michielsen, K.},
   Journal                  = {Physical Review X},
   Year                     = {2020},
   Pages                    = {021002},
