### 2020-10-26 #2675 @kyleniemeyer
@oliviaguest @miriammenzel I have seen behavior like that with combined citations that include the same surname, so I think it is ok and probably not (easily) changeable
